% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-misc.R
\docType{data}
\name{adjnoun}
\alias{adjnoun}
\title{Words in David Copperfield}
\format{
igraph object
}
\source{
http://www-personal.umich.edu/~mejn/netdata/
}
\usage{
adjnoun
}
\description{
A network of common adjective and noun adjacencies for the novel "David Copperfield" by Charles Dickens, as described by M. Newman. Nodes represent the most commonly occurring adjectives and nouns in the book. Edges connect any pair of words that occur in adjacent position in the text of the book.
}
\references{
Newman, Mark EJ. "Finding community structure in networks using the eigenvectors of matrices." \emph{Physical Review E} 74.3 (2006): 036104.
}
\keyword{datasets}
