% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ally}
\alias{ally}
\title{A list of 10 time periods of data on alliance ties between 50 countries}
\format{
An object of class \code{list} of length 10.
}
\usage{
ally
}
\description{
A list of 10 time periods of data on alliance ties between 50 countries
}
\keyword{datasets}
