% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMnet01}
\alias{AMnet01}
\title{Co-sponsorship network pre-redistricting}
\format{
A square binary matrix.
}
\usage{
data(AMnet01)
}
\description{
A binary adjacency matrix representing co-sponsorship ties among legislators in 2001, before redistricting.
}
\keyword{datasets}
