% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMnet03}
\alias{AMnet03}
\title{Co-sponsorship network post-redistricting}
\format{
An object of class \code{network} of length 5.
}
\usage{
AMnet03
}
\description{
Co-sponsorship network post-redistricting
}
\keyword{datasets}
