% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_10}
\alias{animal_10}
\title{Philetairus socius}
\format{list of igraph objects}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_10
}
\description{
Species: \emph{Philetairus socius}

Taxonomic class: Aves

Population type: free-ranging

Geographical location: Kimberley, South Africa

Data collection technique: mark recapture

Interaction type: social projection bipartite

Definition of interaction: A network edge was drawn between individuals that used the same nest chambers either for roosting or nest-building at any given time within a series of observations at the same colony in the same year, either together in the nest chamber at the same time or at different times.

Edge weight type: unweighted

Total duration of data collection: 10 months

Time resolution of data collection (within a day):

Time span of data collection (within a day): focal follow/ad libitum

Note: Networks represent social data collected from 23 colonies of sociable weavers
}
\references{
Dijk, René E., et al. "Cooperative investment in  public goods is kin directed in communal  nests of social birds." Ecology letters 17.9  (2014):  1141-1148.
}
\keyword{datasets}
