% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_12}
\alias{animal_12}
\title{Ants Proximity (weighted)}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_12
}
\description{
Species: \emph{Camponotus fellah}

Taxonomic class: Insecta

Population type: captive

Geographical location: University of Lausanne, Laussane, Switzerland

Data collection technique: video

Interaction type: physical contact

Definition of interaction: A pair of ants was considered to interact when the front end of one ant was located within the trapezoidal shape representing the other ant.

Edge weight type: frequency

Total duration of data collection: 1day

Time resolution of data collection (within a day): 0.5 sec

Time span of data collection (within a day): 24 hours

Note: Networks represent six separate colonies of the ant. The authors recorded the position and orientation of all individuals twice per second to reconstruct spatial movement and infer all social interactions occurring over the 41 days of the experiment.
}
\references{
Mersch, Danielle P., Alessandro Crespi, and Laurent Keller.  "Tracking individuals shows spatial fidelity is a  key regulator of ant social organization."Science 340.6136  (2013):  1090-1093.
}
\keyword{datasets}
