% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_13}
\alias{animal_13}
\title{Ants Trophallaxis (weighted)}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_13
}
\description{
Species: \emph{Camponotus pennsylvanicus}

Taxonomic class: Insecta

Population type: captive

Geographical location: Old Main, State College, Pennsylvania

Data collection technique: video

Interaction type: trophallaxis

Definition of interaction: "

A trophallaxis event was recorded when ants engaged in mandible-to-mandible contact for greater than 1 s"

Edge weight type: duration

Total duration of data collection: 1 day

Time resolution of data collection (within a day): 1sec

Time span of data collection (within a day): 20min

Note: Networks represent two C. pennsylvanicus colonies. Each colony was filmed for approximately 30 minutes for 8 consecutive nights.
}
\references{
Quevillon, Lauren E., et al. "Social, spatial, and  temporal organization in a complex insect society."  Scientific reports 5 (2015).
}
\keyword{datasets}
