% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_15}
\alias{animal_15}
\title{Asianelephants Dominance (unweighted)}
\format{list of igraph objects}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_15
}
\description{
Species: \emph{Elephas maximus}

Taxonomic class: Mammalia

Population type: free-ranging

Geographical location: Uda Walawe National Park, Sri Lanka

Data collection technique: focal sampling

Interaction type: dominance

Definition of interaction: Indicators of dominance as well as subordination was included. If a series of interactions occurred during a particular event, the winners/losers were determined only on conclusion of the event, when individuals or groups moved apart.

Edge weight type: unweighted

Total duration of data collection: 206days

Time resolution of data collection (within a day): 1sec

Time span of data collection (within a day): 5.5 hours

Note:
}
\references{
de Silva, Shermin, Volker Schmid, and George Wittemyer.  "Fission–fusion processes weaken dominance networks of female  Asian elephants in a productive habitat." Behavioral  Ecology (2016):  arw153.
}
\keyword{datasets}
