% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_17}
\alias{animal_17}
\title{Bats Foodsharing (weighted)}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_17
}
\description{
Species: \emph{Desmodus rotundus}

Taxonomic class: Mammalia

Population type: captive

Geographical location: Bloomfield Hills, Michigan, USA

Data collection technique: video

Interaction type: trophallaxis

Definition of interaction: Food sharing event were defined as periods where food could be passed that lasted at least 5 s and were separated by more than 5s.

Edge weight type: duration

Total duration of data collection: 2 hours

Time resolution of data collection (within a day): 1 sec

Time span of data collection (within a day): 2 hours

Note:
}
\references{
Carter, Gerald G., and Gerald S. Wilkinson. "Food  sharing in vampire bats:  reciprocal help predicts  donations more than relatedness or harassment."Proceedings of  the Royal Society of London B:  Biological  Sciences 280.1753 (2013):  20122573.
}
\keyword{datasets}
