% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_18}
\alias{animal_18}
\title{Bats Roostuse (weighted)}
\format{list of igraph objects}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_18
}
\description{
Species: \emph{Myotis sodalis}

Taxonomic class: Mammalia

Population type: free-ranging

Geographical location: Pickaway County, Ohio, USA

Data collection technique: radio tags

Interaction type: social projection bipartite

Definition of interaction: Roost network was first contructed as a two-mode network that consisted of bats and roosts. Single-mode projection of the bat nodes was used to assess colony social structure.

Edge weight type: unweighted

Total duration of data collection: 1 year

Time resolution of data collection (within a day):

Time span of data collection (within a day): 8 hours

Note:
}
\references{
Silvis, Alexander, et al. "Roosting and foraging social  structure of the endangered Indiana bat (Myotis  sodalis)." PloS one 9.5 (2014):  e96937.
}
\keyword{datasets}
