% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_19}
\alias{animal_19}
\title{Bison Dominance (weighted)}
\format{list of igraph objects}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_19
}
\description{
Species: \emph{Bison bison}

Taxonomic class: Mammalia

Population type: semi-ranging

Geographical location: Moiese, Montana, USA

Data collection technique: survey scan

Interaction type: dominance

Definition of interaction: Only those aggressive interactions were analyzed that led to an outcome in which the lost and the winner could be categorized unambiguosly

Edge weight type: frequency

Total duration of data collection: 2 months

Time resolution of data collection (within a day): 1sec

Time span of data collection (within a day): focal follow/ad libitum

Note:
}
\references{
Dale F Lott. Dominance relations and breeding rate  in mature male American bison. Zeitschrift für  Tierpsychologie, 49(4): 418-432, 1979.
}
\keyword{datasets}
