% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_2}
\alias{animal_2}
\title{Guppy Proximity Frequency (weighted)}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_2
}
\description{
Species: \emph{Poecilia reticulata}

Taxonomic class: Actinopterygii

Population type: captive

Geographical location: Louisville, Kentucky

Data collection technique: video

Interaction type: spatial proximity

Definition of interaction: A contact phase between two individuals was one or more consecutive observations in which they were recorded as being within four body length to one another.

Edge weight type: frequency

Total duration of data collection: 90min

Time resolution of data collection (within a day): 10 seconds

Time span of data collection (within a day): 90min

Note: Networks represent 10 independed replicates of grourp of 5 guppies that were familiar with each other
}
\references{
Hasenjager, Matthew J., and Lee Alan Dugatkin. "Familiarity  affects network structure and information flow in  guppy (Poecilia reticulata) shoals." Behavioral Ecology (2016):   arw152.
}
\keyword{datasets}
