% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_20}
\alias{animal_20}
\title{Cattle Dominance (weighted)}
\format{list of igraph objects}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_20
}
\description{
Species: \emph{Cattle N/A}

Taxonomic class: Mammalia

Population type: semi-ranging

Geographical location: Jeanerette, Louisiana, USA

Data collection technique: survey scan

Interaction type: dominance

Definition of interaction: The dominance relationship was ascertained from direct physical contests.

Edge weight type: frequency

Total duration of data collection: Not specified

Time resolution of data collection (within a day): 1 sec

Time span of data collection (within a day): 1 hour

Note:
}
\references{
Martin W. Schein and Milton H. Fohrman. Social  dominance relationships in a herd of dairy  cattle. The British J. of Animal Behaviour,  3(2): 45-55, 1955.
}
\keyword{datasets}
