% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_21}
\alias{animal_21}
\title{Dolphin Association (weighted)}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_21
}
\description{
Species: \emph{Tursiops truncatus}

Taxonomic class: Mammalia

Population type: free-ranging

Geographical location: Cedar Key, Florida, USA

Data collection technique: survey scan

Interaction type: spatial proximity

Definition of interaction: Interactions characterized by prey capture or persistent incidents of prey searching as indicated by long dives or specialized feeding behaviours with direction shifts between surfacings.

Edge weight type: frequency

Total duration of data collection: 124 days

Time resolution of data collection (within a day):

Time span of data collection (within a day): 5min

Note: Four networks:  an overall network that does not take behaviour into account, and the socialize network, the travel network and the forage network that correspond to their respective behaviours.
}
\references{
Gazda, Stefanie, et al. "The importance of delineating  networks by activity type in bottlenose dolphins  (Tursiops truncatus) in Cedar Key, Florida." Royal  Society open science 2.3 (2015):  140263.
}
\keyword{datasets}
