% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_22}
\alias{animal_22}
\title{Dolphin Proximity (unweighted)}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_22
}
\description{
Species: \emph{Tursiops truncatus}

Taxonomic class: Mammalia

Population type: free-ranging

Geographical location: Fiordland, New Zealand

Data collection technique: survey scan

Interaction type: spatial proximity

Definition of interaction: All members of a school were assumed associated. Half-weight index (HWI) was used to quantify the frequency of association among individuals.

Edge weight type: frequency

Total duration of data collection: 594 days

Time resolution of data collection (within a day):

Time span of data collection (within a day): Not specified

Note:
}
\references{
Lusseau, David, et al. "The bottlenose dolphin community  of Doubtful Sound features a large proportion  of long-lasting associations." Behavioral Ecology and Sociobiology  54.4 (2003):  396-405.
}
\keyword{datasets}
