% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_23}
\alias{animal_23}
\title{Elephantseal Dominance (weighted)}
\format{list of igraph objects}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_23
}
\description{
Species: \emph{Mirounga angustirostris}

Taxonomic class: Mammalia

Population type: free-ranging

Geographical location: California, USA

Data collection technique: survey scan

Interaction type: dominance

Definition of interaction: Dominance status determined by Elo rating based on winner of an competitive interaction and intensity of interaction

Edge weight type: frequency

Total duration of data collection: 69 days

Time resolution of data collection (within a day):

Time span of data collection (within a day): focal follow/ad libitum

Note: Network represent interactions collected over different years and/or in different colonies
}
\references{
Casey, Caroline, et al. "Rival assessment among northern  elephant seals:  evidence of associative learning during  male–male contests." Royal Society open science 2.8  (2015):  150228.
}
\keyword{datasets}
