% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_24}
\alias{animal_24}
\title{Hyenas Groupmembership (weighted)}
\format{list of igraph objects}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_24
}
\description{
Species: \emph{Crocuta crocuta}

Taxonomic class: Mammalia

Population type: free-ranging

Geographical location: Masai Mara National Reserve, Kenya

Data collection technique: survey scan

Interaction type: group membership

Definition of interaction: Association patterns were recorded based on the co-occurrence of each pair of individuals, during the period for which they were concurrently present in the clan.

Edge weight type: twice_weight_index

Total duration of data collection: 4 months (approx 16 days each month)

Time resolution of data collection (within a day): 15min

Time span of data collection (within a day): few hours

Note: The three social networks were collected during periods of low (networks A and C) and high (network B) prey abundance.
}
\references{
Holekamp, Kay E., et al. "Society, demography and  genetic structure in the spotted hyena." Molecular  Ecology 21.3 (2012):  613-632.
}
\keyword{datasets}
