% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_25}
\alias{animal_25}
\title{Kangeroos Proximity (weighted)}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_25
}
\description{
Species: \emph{Macropus giganteus}

Taxonomic class: Mammalia

Population type: free-ranging

Geographical location: New South Wales, Australia

Data collection technique: survey scan

Interaction type: spatial proximity

Definition of interaction: Two individuals were assumed to be associating if they occured within 120 cm of another at set 15-min intervals in the enclosure.

Edge weight type: frequency

Total duration of data collection: Not specified

Time resolution of data collection (within a day):

Time span of data collection (within a day): focal follow/ad libitum

Note:
}
\references{
TR Grant. Dominance and association among members of  a captive and a free-ranging group of  grey kangaroos (Macropus giganteus). Animal Behaviour, 21(3): 449-456,  1973.
}
\keyword{datasets}
