% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_29}
\alias{animal_29}
\title{Rhesusmacaque Association (weighted)}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_29
}
\description{
Species: \emph{Macaca mulatta}

Taxonomic class: Mammalia

Population type: captive

Geographical location: Rijswijk, the Netherlands

Data collection technique: survey scan

Interaction type: physical contact

Definition of interaction: Edge defined based on individuals sitting in contact with eah other.

Edge weight type: frequency

Total duration of data collection: 3years

Time resolution of data collection (within a day): 1 hour

Time span of data collection (within a day): few minutes

Note: Two networks represent proximity (contact sits) and grooming patterns - who was sitting in contact with whom and who was grooming whom.
}
\references{
Massen, Jorg JM, and Elisabeth HM Sterck. "Stability  and durability of intra-and intersex social bonds  of captive rhesus macaques (Macaca mulatta)." International  Journal of Primatology 34.4 (2013):  770-791.
}
\keyword{datasets}
