% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_3}
\alias{animal_3}
\title{Hirundo rustica}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_3
}
\description{
Species: \emph{Hirundo rustica}

Taxonomic class: Aves

Population type: free-ranging

Geographical location: Boulder County, Colorado, USA

Data collection technique: logger

Interaction type: physical contact

Definition of interaction: Interaction 0.1 m and closer

Edge weight type: frequency

Total duration of data collection: 11 days

Time resolution of data collection (within a day): 1 sec

Time span of data collection (within a day): 6 hours

Note: Two networks were constructed with edges weighted by the number of interactions at two spatial proximities:  body contact interactions and all other spatially proximate interactions
}
\references{
Levin, Iris I., et al. "Stress response, gut  microbial diversity and sexual signals correlate with  social interactions." Biology Letters 12.6 (2016):  20160352.
}
\keyword{datasets}
