% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_30}
\alias{animal_30}
\title{Rhesusmacaque Dominance (weighted)}
\format{list of igraph objects}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_30
}
\description{
Species: \emph{Macaca fuscata}

Taxonomic class: Mammalia

Population type: semi-ranging

Geographical location: Arashiyama, Japan

Data collection technique: survey scan

Interaction type: dominance

Definition of interaction: The dominance relations between females were determined based on approach-retreat episodes around the food. The dominance range order was arranged based on these dyadic relations.

Edge weight type: frequency

Total duration of data collection: 6 months

Time resolution of data collection (within a day): 2 hours

Time span of data collection (within a day): focal follow/ad libitum

Note:
}
\references{
Takahata, Yukio. "Diachronic changes in the dominance relations  of adult female Japanese monkeys of the  Arashiyama B group." The monkeys of Arashiyama.  State University of New York Press, Albany  (1991):  123-139.
}
\keyword{datasets}
