% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_31}
\alias{animal_31}
\title{Sheep Dominance (weighted)}
\format{list of igraph objects}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_31
}
\description{
Species: \emph{Ovis canadensis}

Taxonomic class: Mammalia

Population type: semi-ranging

Geographical location: Montana, USA

Data collection technique: focal sampling

Interaction type: dominance

Definition of interaction: Social status was determined by assembling a win-loss matrix based on the outcome of agonistic interactions. The winner of a dominance fight, involving a series of butts and clashes, was recorded as winning one interaction.

Edge weight type: frequency

Total duration of data collection: 15 months

Time resolution of data collection (within a day):

Time span of data collection (within a day): focal follow/ad libitum

Note:
}
\references{
Christine C Hass. Social status in female bighorn  sheep (Ovis canadensis):  Expression, development and reproductive  correlates. J. of Zoology, 225(3): 509-523, 1991.
}
\keyword{datasets}
