% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_32}
\alias{animal_32}
\title{Spidermonkeys Contact (weighted)}
\format{list of igraph objects}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_32
}
\description{
Species: \emph{Ateles hybridus}

Taxonomic class: Mammalia

Population type: free-ranging

Geographical location: Colombia

Data collection technique: focal sampling

Interaction type: physical contact

Definition of interaction: Physical contact event between the two individuals (e.g. grooming, mating and embracing)

Edge weight type: duration

Total duration of data collection: 2 years

Time resolution of data collection (within a day): 15min

Time span of data collection (within a day): focal follow/ad libitum

Note:
}
\references{
Rimbach, Rebecca, et al. "Brown spider monkeys (Ateles  hybridus):  a model for differentiating the role  of social networks and physical contact on  parasite transmission dynamics." Phil. Trans. R. Soc.  B 370.1669 (2015):  20140110.
}
\keyword{datasets}
