% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_34}
\alias{animal_34}
\title{Zebra Groupmembership (weighted)}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_34
}
\description{
Species: \emph{Equus grevyi}

Taxonomic class: Mammalia

Population type: free-ranging

Geographical location: Kenya

Data collection technique: survey scan

Interaction type: group membership

Definition of interaction: A group was defined as a set of one or more individuals that is spatially cohesive and distinct from other groups at the time of observation. Edges were constructed based on half-weight index (HWI).

Edge weight type: unweighted

Total duration of data collection: 3 months

Time resolution of data collection (within a day):

Time span of data collection (within a day): focal follow/ad libitum

Note:
}
\references{
Siva R Sundaresan, Ilya R Fischhoff, Jonathan Dushoff,  and Daniel I Rubenstein. Network metrics reveal  differences in social organization between two fission-fusion  species, Grevy's zebra and onager. Oecologia, 151(1): 140-149,  2007.
}
\keyword{datasets}
