% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_36}
\alias{animal_36}
\title{Lizard Proximity (weighted)}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_36
}
\description{
Species: \emph{Tiliqua rugosa}

Taxonomic class: Reptilia

Population type: free-ranging

Geographical location: Kungara, South Australia

Data collection technique: logger

Interaction type: spatial proximity

Definition of interaction: Two lizards were assumed to had made a social contact if they were within 2 m of each other at any of the synchronized 10 min GPS locations.

Edge weight type: simple_ratio_index

Total duration of data collection: 120 days

Time resolution of data collection (within a day): 10 minutes

Time span of data collection (within a day): 24 hours

Note:
}
\references{
Bull, C. M., S. S. Godfrey, and D.  M. Gordon. "Social networks and the spread  of Salmonella in a sleepy lizard population."  Molecular Ecology 21.17 (2012):  4386-4392.
}
\keyword{datasets}
