% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_4}
\alias{animal_4}
\title{Branta leucopsis}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_4
}
\description{
Species: \emph{Branta leucopsis}

Taxonomic class: Aves

Population type: captive

Geographical location: Heteren, Netherlands

Data collection technique: survey scan

Interaction type: foraging

Definition of interaction: Animals grazing on the same patch during a sampling period were assumed to be associating (gambit of the group)

Edge weight type: frequency

Total duration of data collection: 15 days

Time resolution of data collection (within a day): 4min

Time span of data collection (within a day): 4 hours

Note: Networks represent male and female only network
}
\references{
Kurvers, Ralf HJM, et al. "Contrasting context dependence  of familiarity and kinship in animal social  networks." Animal Behaviour 86.5 (2013):  993-1001.
}
\keyword{datasets}
