% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_5}
\alias{animal_5}
\title{White Leghorn}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_5
}
\description{
Species: \emph{White leghorn}

Taxonomic class: Aves

Population type: captive

Geographical location: Manhattan, Kansas

Data collection technique: mn/unspecified

Interaction type: dominance

Definition of interaction: Dominance interaction recoded based on agonistic behavior until each bird has established its dominance and/or subordination.

Edge weight type: frequency

Total duration of data collection:

Time resolution of data collection (within a day):

Time span of data collection (within a day):

Note:
}
\references{
Guhl, A. M. "Social behavior of the domestic  fowl." Transactions of the Kansas Academy of  Science (1903-) 71.3 (1968):  379-384.
}
\keyword{datasets}
