% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_6}
\alias{animal_6}
\title{Haemorhous mexicanus}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_6
}
\description{
Species: \emph{Haemorhous mexicanus}

Taxonomic class: Aves

Population type: free-ranging

Geographical location: Virginia Tech, Virginia, USA

Data collection technique: RFID

Interaction type: social projection bipartite

Definition of interaction: A machine learning algorithm was applied to identify clusters of detections on feeders. Next, the network was generated based on patterns of co-occurrence by individuals in the same feeding events. Asociations between birds were defined using the simple ratio index.

Edge weight type: simple_ratio_index

Total duration of data collection: 137 days

Time resolution of data collection (within a day): 1 sec

Time span of data collection (within a day): 24 hours

Note:
}
\references{
Adelman, James S., et al. "Feeder use predicts  both acquisition and transmission of a contagious  pathogen in a North American songbird." Proc.  R. Soc. B. Vol. 282. No. 1815.  The Royal Society, 2015.
}
\keyword{datasets}
