% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animal_7}
\alias{animal_7}
\title{Zonotrichia Atricapilla}
\format{
list of igraph objects
}
\source{
https: //bansallab.github.io/asnr/
}
\usage{
animal_7
}
\description{
Species: \emph{Zonotrichia atricapilla}

Taxonomic class: Aves

Population type: free-ranging

Geographical location: California, USA

Data collection technique: survey scan

Interaction type: group membership

Definition of interaction: A flock was defined as a group of birds within an approximately 5-metre radius. Social networks of flock comembership was constructed where nodes represent individual birds and edges represent the simple ratio association index

Edge weight type: simple_ratio_index

Total duration of data collection: 3 months

Time resolution of data collection (within a day):

Time span of data collection (within a day): focal follow/ad libitum

Note: Two networks collected over two consecutive years
}
\references{
Arnberg, Nina N., et al. "Social network structure  in wintering golden‐crowned sparrows is not correlated  with kinship." Molecular ecology 24.19 (2015):  5034-5044.
}
\keyword{datasets}
