% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{ants_1}
\alias{ants_1}
\title{Ant Colony (I)}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data.html#ants
}
\usage{
ants_1
}
\description{
These are observations of ritual dominance activities in an ant community (a collection of 16 female Leptothorax allardycei ants over 18.2 hours in a queenright colony)
}
\references{
B. J. Cole, 1981, "Dominance hierarchies in Leptothorax ants" \emph{Science}, 212: 83-84.
}
\seealso{
\link{ants_2}
}
\keyword{datasets}
