% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{ants_2}
\alias{ants_2}
\title{Ant Colony (II)}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data.html#ants
}
\usage{
ants_2
}
\description{
These are observations of ritual dominance activities in an ant community (a collection of 13 female Leptothorax allardycei ants in a queenless colony)
}
\references{
B. J. Cole, 1981, "Dominance hierarchies in Leptothorax ants" \emph{Science}, 212: 83-84.
}
\seealso{
\link{ants_1}
}
\keyword{datasets}
