% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{arenas_email}
\alias{arenas_email}
\title{Arenas Email}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://deim.urv.cat/~aarenas/data/welcome.htm
}
\usage{
arenas_email
}
\description{
This is the email communication network at the University Rovira i Virgili in Tarragona in the south of Catalonia in Spain. Nodes are users and each edge represents that at least one email was sent.  The direction of emails or the number of emails are not stored.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Roger Guimerà, Leon Danon, Albert Díaz-Guilera, Francesc Giralt, and Alex  Arenas. Self-similar community structure in a network of human interactions. Phys. Rev. E, 68(6):065103, 2003.
}
\keyword{datasets}
