% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{arenas_meta}
\alias{arenas_meta}
\title{Arenas Metabolic}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://deim.urv.cat/~aarenas/data/welcome.htm
}
\usage{
arenas_meta
}
\description{
This is the metabolic network of the roundworm Caenorhabditis elegans.  Nodes are metabolites (e.g., proteins), and edges are interactions between them.  Since a metabolite can iteract with itself, the network contains loops.  The interactions are undirected.  There may be multiple interactions between any two metabolites.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Jordi Duch and Alex Arenas. Community detection in complex networks using extremal optimization. Phys. Rev. E, 72(2):027104, 2005.
}
\keyword{datasets}
