% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-misc.R
\docType{data}
\name{atp}
\alias{atp}
\title{ATP Tennis (1968-2019)}
\format{list of igraph objects}
\source{
Networks constructed from data that was gathered and compiled by Jeff Sackmann (https://github.com/JeffSackmann). Please give credit to him if you use this data.
}
\usage{
atp
}
\description{
The dataset includes all ATP tennis matches from 1968-2019. The networks are directed pointing from the loser to the winner.
Each network contains the following attributes:
\preformatted{
Edge attributes:
   - surface: on which surface the match(es) took place (e.g. "Hard", "Grass", "Clay")
   - weight: number of times Player A lost to Player B on surface X
Vertex attributes:
   - hand: if player is (L)eft or (R)ight handed (or (U)nknown)
   - age: age of player during the season
   - country: home country of player
}
Check out \url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0017249} for a potential use case.
}
\seealso{
\link{wta}
}
\keyword{datasets}
