\name{beach}
\alias{beach}
\docType{data}
\title{Freeman, Freeman, and Michaelson Beach}
\description{
beach: Two 43 by 43 matrices.

\bold{BACKGROUND} This was a study of windsurfers on a beach in southern California during the fall of 1986. The windsurfing community was fairly clearly divided into at least two sub-communities. Members of each community seemed, to some degree, to limit their interaction to fellow group members. Contacts between members of the two groups occurred, but these were less frequent. Observations of 43 individuals were made for 31 days. All interpersonal contacts among collections of these individuals were recorded. Then all 43 individuals were interviewed following the end of observation. Data on each individual's perception of social affiliations were collected.

The perceptual data were generated by asking each subject to perform a sequence of card sorting tasks that assigned an index of the perceived closeness of every individual on the beach to each of the other individuals.


}
\usage{data(beach)}

\details{
bb: symmetric, valued, observed interaction.\cr
bc: symmetric, valued, judged closeness.
}
\source{
L. C. Freeman, S. C. Freeman and A. G. Michaelson \emph{On Human Social Intelligence.} Journal of Social and Biological Structures, 11, 1988, 415-425.
}
\references{
L. C. Freeman, S. C. Freeman and A. G. Michaelson \emph{How Humans See Social Groups: A Test of the Sailer-Gaulin Models.} Journal of Quantitative Anthropology, 1, 1989, 229-238.
}
\examples{
data(beach)

##Plot
plot(beach$bb)

##Valued adjacency matrix
as.sociomatrix(beach$bb,attr="edgevalue")


}
\keyword{datasets}
