% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{bible}
\alias{bible}
\title{Nouns in the King's James Bible}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data#bible
}
\usage{
bible
}
\description{
Christoph Romhild recorded 1773 proper nouns--people and places--in the King James Bible. He tallied 63,779 occasions in which pairs of these proper nouns appeared in the same verse in the bible. Many of these, of course, appeared more than once. So the data presented here are tallies, for each pair of proper nouns, of the number of verses in which they appeared together. Romhild worked with Chris Harrison, and together, they produced some elegant visual images of the data. They are displayed in the source listed below.
}
\references{
http://chrisharrison.net/projects/bibleviz/index.html
}
\keyword{datasets}
