% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{bkfrac}
\alias{bkfrac}
\title{Bernard/Killworth - Fraternity (rankings)}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data.html#bkfrat
}
\usage{
bkfrac
}
\description{
These data concern interactions among students living in a fraternity at a West Virginia college. All subjects had been residents in the fraternity from three months to three years. BKFRAB records the number of times a pair of subjects were seen in conversation by an "unobtrusive" observer (who walked through the public areas of the building every fifteen minutes, 21 hours a day, for five days). BKFRAC contains rankings made by the subjects of how frequently they interacted with other subjects in the observation week.
}
\references{
#' Bernard H. R., Killworth P. and Sailer L. (1980). Informant accuracy in social network data IV. \emph{Social Networks}, 2, 191-218.

Bernard H. R., Killworth P. and Sailer L. (1982). Informant accuracy in social network data V. \emph{Social Science Research}, 11, 30-66.

Romney A. K. and Weller S. (1984). Predicting informant accuracy from patterns of recall among individuals. \emph{Social Networks}, 6, 59-78.
}
\seealso{
\link{bkfrab}
}
\keyword{datasets}
