% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{bkoffb}
\alias{bkoffb}
\title{Bernard/Killworth - Office (interaction)}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data.html#bkoff
}
\usage{
bkoffb
}
\description{
These data concern interactions in a small business office, recorded by an "unobtrusive" observer. Observations were made as the observer patrolled a fixed route through the office every fifteen minutes during two four-day periods. BKOFFB contains the observed frequency of interactions; BKOFFC contains rankings of interaction frequency as recalled by the employees over the two-week period.
}
\references{
Bernard H. R., Killworth P. and Sailer L. (1980). Informant accuracy in social network data IV. \emph{Social Networks}, 2, 191-218.

Bernard H. R., Killworth P. and Sailer L. (1982). Informant accuracy in social network data V. \emph{Social Science Research}, 11, 30-66.

Romney A. K. and Weller S. (1984). Predicting informant accuracy from patterns of recall among individuals. \emph{Social Networks}, 6, 59-78.
}
\seealso{
\link{bkoffc}
}
\keyword{datasets}
