% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{bktecb}
\alias{bktecb}
\title{Bernard/Killworth - Tech company (interaction)}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data.html#bktec
}
\usage{
bktecb
}
\description{
These data concern interactions in a technical research group at a West Virginia university. BKTECB contains a frequency record of interactions, made by an observer every half-hour during one five-day work week. BKTECC contains the personal rankings of the remembered frequency of interactions in the same period.
}
\references{
Bernard H. R., Killworth P. and Sailer L. (1980). Informant accuracy in social network data IV. \emph{Social Networks}, 2, 191-218.

Bernard H. R., Killworth P. and Sailer L. (1982). Informant accuracy in social network data V. \emph{Social Science Research}, 11, 30-66.

Romney A. K. and Weller S. (1984). Predicting informant accuracy from patterns of recall among individuals. \emph{Social Networks}, 6, 59-78.
}
\seealso{
\link{bktecc}
}
\keyword{datasets}
