% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{bott}
\alias{bott}
\title{Preschool}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data.html#bott
}
\usage{
bott
}
\description{
The data were collected in 1926 in a preschool in Toronto. Observations were made on each child in turn who was defined as a "focal" individual. Instances in which the focal child (1) talked to another, (2) interfered with another, (3) watched another, (4) imitated another or (5) cooperated with another were tabulated along with the name of the other to whom the social behavior was directed. The result was tabulated in five matrices.
}
\details{
The five different relations are given in the relation edge attribute.
}
\references{
Bott, H. "Observations of play activities in a nursery school," \emph{Genetic Psychology Monographs}, 1928, 4: 44-88.
}
\keyword{datasets}
