% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{brunson_club_membership}
\alias{brunson_club_membership}
\title{Brunson Club Membership}
\format{
igraph object
}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from https://github.com/corybrunson/triadic
}
\usage{
brunson_club_membership
}
\description{
This bipartite network contains membership information of corporate executive officers in social organisations such as clubs and boards. Left nodes represent persons and right nodes represent social organisations. An edge between a person and a social organization shows that the person has a memberstatus.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Katherine Faust. Centrality in affiliation networks. Social Networks, 19(2):157--191, 1997.
}
\keyword{datasets}
