% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{brunson_corporate_leadership}
\alias{brunson_corporate_leadership}
\title{Brunson Corporate Leadership}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from https://github.com/corybrunson/triadic
}
\usage{
brunson_corporate_leadership
}
\description{
This bipartite network contains person–company leadership information between companies and 20 corporate directors. The data was collected in 1962. Left nodes represent persons and right nodes represent companies. An edge between a person and a company shows that the person had a leadership position in that company.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Roy Barnes and Tracy Burkett. Structural redundancy and multiplicity in corporate networks. International Network for Social Network Analysis, 30(2), 2010.
}
\keyword{datasets}
