% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{brunson_revolution}
\alias{brunson_revolution}
\title{Brunson Revolution}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from https://github.com/corybrunson/triadic
}
\usage{
brunson_revolution
}
\description{
This bipartite network contains membership information of 136 people in 5 organisations dating back to the time before the American Revolution. The list includes well-known people such as the American activist Paul Revere. Left nodes represent persons and right nodes represent organisations. An edge between a person and an organization shows that the person was a member of the organisation.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

NA
}
\keyword{datasets}
