% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{brunson_south_africa}
\alias{brunson_south_africa}
\title{Brunson South Africa}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from https://github.com/corybrunson/triadic
}
\usage{
brunson_south_africa
}
\description{
This bipartite network contains person–company shared leadership relations of "the five most representative companies" that are claimed to represent "the small inner ring of South African Finance". Left nodes represent persons and right nodes represent companies. An edge between a person and a company shows that the person had a leadership position in that company.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

John Atkinson Hobson. The Evolution of Modern Capitalism: A Study of Machine  Production. The Walter Scott publishing co., ltd., 1919.
}
\keyword{datasets}
