% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{ceos_clubs}
\alias{ceos_clubs}
\title{Ceo's and Clubs}
\format{(bipartite) igraph object}
\source{
http://moreno.ss.uci.edu/data.html#galas
}
\usage{
ceos_clubs
}
\description{
These data give the affiliation network of 26 CEO's of major corporations and banks and their spouses to 15 clubs, corporate and cultural boards.. Data were collected in the Minneapolis area. Membership was during the period 1978-1981.
}
\references{
Galaskiewicz J (1985). Social Organization of an Urban Grants Economy. New York. Academic Press.
}
\keyword{datasets}
