% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{chicagoroad}
\alias{chicagoroad}
\title{Road Transportation Network Chicago}
\format{
igraph object
}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://www.bgu.ac.il/~bargera/tntp/
}
\usage{
chicagoroad
}
\description{
This is the road transportation network of the Chicago region (USA).  Nodes are transport nodes, and edges are connections.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

R. W. Eash, K. S. Chon, Y. J. Lee, and D. E. Boyce. Equilibrium traffic assignment on an aggregated highway network for  sketch planning. Transportation Research Record, 994:30--37, 1983.
}
\keyword{datasets}
