% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-misc.R
\docType{data}
\name{clique_graph}
\alias{clique_graph}
\title{Illustrating cliques}
\format{
igraph object
}
\usage{
clique_graph
}
\description{
A small graph to illustrate the concept of cliques and k-cores.
}
\keyword{datasets}
