% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cname}
\alias{cname}
\title{a vector of country names}
\format{
An object of class \code{character} of length 195.
}
\usage{
cname
}
\description{
a vector of country names
}
\keyword{datasets}
