% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-misc.R
\docType{data}
\name{coleman}
\alias{coleman}
\title{Coleman's High School Friendship Data}
\format{
list of two igraph objects
}
\usage{
coleman
}
\description{
James Coleman (1964) reports research on self-reported friendship ties among 73 boys in a small high school in Illinois over the 1957-1958 academic year. Networks of reported ties for all 73 informants are provided for two time points (fall and spring).
}
\references{
Coleman, J. S. (1964). Introduction to Mathematical Sociology. New York: Free Press.
}
\keyword{datasets}
