% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONT}
\alias{CONT}
\title{a contiguity matrix, whose elements equal to 1 if two countries are contiguous (share a land border)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 195 rows and 195 columns.
}
\usage{
CONT
}
\description{
a contiguity matrix, whose elements equal to 1 if two countries are contiguous (share a land border)
}
\keyword{datasets}
