% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contig}
\alias{contig}
\title{A list of 8 time periods of data on contiguity between 134 countries}
\format{
An object of class \code{list} of length 8.
}
\usage{
contig
}
\description{
A list of 8 time periods of data on contiguity between 134 countries
}
\keyword{datasets}
