% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contiguity}
\alias{contiguity}
\title{A list of 10 time periods of data on contiguity between 50 countries}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 50 rows and 50 columns.
}
\usage{
contiguity
}
\description{
A list of 10 time periods of data on contiguity between 50 countries
}
\keyword{datasets}
