% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-misc.R
\docType{data}
\name{core_graph}
\alias{core_graph}
\title{Illustrating core-periphery}
\format{
igraph object
}
\usage{
core_graph
}
\description{
A graph to illustrate the concept of core-periphery.
}
\keyword{datasets}
