% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-misc.R
\docType{data}
\name{cosponsor}
\alias{cosponsor}
\title{Senat 2015 Bill cosponsorship}
\format{
two-mode network as igraph object
}
\usage{
cosponsor
}
\description{
Bill cosponsorship network for the 115th Senate obtained from govtrack.us
}
\keyword{datasets}
