% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_11}
\alias{covert_11}
\title{Christmas Eve Bombings Indonesia 2000}
\format{list of igraph objects}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks) or here http://doitapps.jjay.cuny.edu/jjatt/data.php
}
\usage{
covert_11
}
\description{
This is a time series that treat specific attacks as endpoints and depict the evolution of relations between individuals indirectly and directly associated with the 2000 Christmas Eve bombing.  http://en.wikipedia.org/wiki/Christmas_Eve_2000_Indonesia_bombings
1-mode stacked matrices 27 x 27 person by person. Data for 11 time periods plus kinship data. Undirected, valued ties.
Tie value codes for kinship matrix:
0 = No Kinship // 1 = In-laws // 2 = Cousins // 3 = Sibling // 4 =
Parent/Child // 5 = Married // 6 = Grandparent/Child // 7 = Significant
Other
Tie value codes for time series matrices:
0 = No
relation // 1 = Acquaintances/distant family ties (interaction limited
to radical organisation activities) // 2 = Friends/Moderately close
family (inc co-workers/ roommates) Operational/Org
leadership/Operational lies (e.g. worked closely on a bombing together)
// 3 = Close friends/family, tight-knit operational cliques
}
\keyword{datasets}
