% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_13}
\alias{covert_13}
\title{Cocaine Dealing Natarajan}
\format{igraph object}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Reconstructed from Natarajan, M. "Understanding the structure of a drug trafficking organization: a conversational analysis." Crime Prevention Studies 11 (2000): 273-298.
}
\usage{
covert_13
}
\description{
This dataset comes from an investigation into to a large cocaine trafficking organization in New York City.
1-mode matrix 28 x 28 persons by persons.
Directed valued relations are communications exchanges / flows of information. Data come from police wiretappings (transcripts of 151 telephone conversations).
}
\references{
Natarajan, M. "Understanding the structure of a drug trafficking organization: a conversational analysis." Crime Prevention Studies 11 (2000): 273-298.
}
\keyword{datasets}
