% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_17}
\alias{covert_17}
\title{Drugnet}
\format{igraph object}
\source{
https://sites.google.com/site/sfeverton18/research/cohesion-and-clustering
}
\usage{
covert_17
}
\description{
This is a dichotomous adjacency matrix of drug users in Hartford. Ties are directed and represent acquaintanceship. The network is a result of two years of ethnographic observations of people's drug habits.
NETWORK  1-mode matrix 293x293 person by person, directed ties. Relations are acquaintanceship.
Attribute dataset includes ethnicity, gender.
Ethnicity codes: 2 = African American; 3 = Puerto Rican/Latino; 1, 5, 6, 7 = white or
other
Gender codes: 1 = male; 2 = female; 0 = unknown
}
\references{
WEEKS, M. R., CLAIR, S., BORGATTI, S. P., RADDA, K. & SCHENSUL, J. J. 2002. Social networks of drug users in high-risk sites: Finding the connections. AIDS and Behaviour, 6, 193-206.
}
\keyword{datasets}
