% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_18}
\alias{covert_18}
\title{FIFA}
\format{list of igraph objects}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Reconstructed by Gemma Edwards from Select Committee 2006 activity report.
}
\usage{
covert_18
}
\description{
Two Networks of Standing Committee membership. These are overt networks with covert elements.
2-Mode persons to Standing Committees (converted to 1-Mode)
340 x 340 persons by persons undirected binary
450 x 450 persons by persons undirected binary
}
\keyword{datasets}
