% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_19}
\alias{covert_19}
\title{Global Suicide Attacks}
\format{
list of igraph objects
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Reconstructed from Benjamin Acosta & Steven J. Childs (2013) ‘Illuminating the Global Suicide-Attack Network’, Studies in Conflict & Terrorism, 36:1, 49-76
}
\usage{
covert_19
}
\description{
Data is on militant organizations between 1985 and 2006. Each node signifies a militant organization or other type of entity that conducts suicide attacks.
1-mode matrix for each year, organization by organization. Undirected, binary ties represent a known physical relationship between agents from different but “connected” organizations.
\preformatted{
1985 10 x 10
1986 4 x 4
1990 4 x 4
1993 4 x 4
1994 4 x 4
1995 9 x 9
1996 6 x6
1997 4 x 4
1998 10 x 10
1999 7 x 7
2000 10 x 10
2001 15 x 15
2002 11 x 11
2003 21 x 21
2004 25 x 25
2005 27 x 27
2006 31 x 31
}
}
\references{
Benjamin Acosta & Steven J. Childs (2013) ‘Illuminating the Global Suicide-Attack Network’, \emph{Studies in Conflict & Terrorism}, 36:1, 49-76
}
\keyword{datasets}
