% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_2}
\alias{covert_2}
\title{9/11 Hijackers}
\format{
igraph object
}
\source{
Reconstructed at Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks).

http://orgnet.com/tnet.html

http://firstmonday.org/ojs/index.php/fm/article/view/941/863#fig4
}
\usage{
covert_2
}
\description{
Famous dataset of the terrorists involved in the 9/11 bombing of the World Trade Centres in 2011. Data was extracted from news reports and ties range from ‘at school with’ to ‘on same plane’.
1-mode matrix 19 x 19 person by person of trusted prior contacts and 1-mode matrix 61 x 61 of other associates.
Ties are undirected and binary. Relations are a mix of prior-contacts like trained together, lived together, financial transactions, at school with, on same flight.
}
\references{
Krebs, Valdis E. "Mapping networks of terrorist cells." \emph{Connections} 24.3 (2002): 43-52.
}
\keyword{datasets}
