% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_20}
\alias{covert_20}
\title{Hamburg Cell 9/11 2001}
\format{
list of igraph objects
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks) or here http://doitapps.jjay.cuny.edu/jjatt/data.php
}
\usage{
covert_20
}
\description{
This is a time series that treat specific attacks as endpoints and depict the evolution of relations between individuals indirectly and directly associated with the sleeper Al Qaeda cell in Hamburg around the time of the 9/11 bombings. http://en.wikipedia.org/wiki/Hamburg_cell
1-mode stacked matrices 35 x 35 person by person, data for 11 time points. Ties are undirected and valued. Tie codes: 0 = No relation // 1 = Acquaintances/distant
family ties (interaction limited to radical organisation activities) // 2
= Friends/Moderately close family (inc co-workers/ roommates)
Operational/Org leadership/Operational lies (e.g. worked closely on a
bombing together) // 3 = Close friends/family, tight-knit operational
cliques.
}
\keyword{datasets}
