% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_22}
\alias{covert_22}
\title{Islamic State Allegiances}
\format{igraph object}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks)
}
\usage{
covert_22
}
\description{
2-mode dataset describing groups allied to
Islamic State and the countries in which they are operating
2-mode matrix 47 x 20 organizations by state, undirected binary ties.
}
\keyword{datasets}
