% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_24}
\alias{covert_24}
\title{Italian Gangs}
\format{igraph object}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks)
}
\usage{
covert_24
}
\description{
Describes Italian gang members and their nationalities. No further contextual data available.
1-Mode matrix 67x67 person by person, relations are co-membership of gangs.
Attribute data is gang member’s country of origin, coded numerically. No codebook available.
}
\keyword{datasets}
