% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_25}
\alias{covert_25}
\title{Jakarta Bombing 2009 / Noordin Top}
\format{
list of igraph objects
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks) or here http://doitapps.jjay.cuny.edu/jjatt/data.php
}
\usage{
covert_25
}
\description{
This is a time series that treat specific attacks as endpoints and depict the evolution of relations between individuals indirectly and directly associated with the 2009 Jakarta bombing. http://en.wikipedia.org/wiki/2009_Jakarta_bombings
This network draws on the same terrorist activities as the Noordin Top network.
1-mode stacked matrices, 28 x 28 person by person, data for kinship, pre-attack and post-attack. Ties undirected and valued.
Codebook available here http://doitapps.jjay.cuny.edu/jjatt/files/Relations_Codebook_Public_Version2.pdf
}
\keyword{datasets}
