% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_27}
\alias{covert_27}
\title{Linux Terrorists}
\format{
list of igraph objects
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks).
}
\usage{
covert_27
}
\description{
These are two datasets about a set of terrorists and the attacks they carry out collected by the MIND Lab at University of Maryland (UMD) (http://www.mindswap.org/).
NETWORK  1-mode matrix 645 x 645 organization by organization, directed ties. Ties are co-location of terrorist attacks.
1-mode matrix 260 x 260 organization by organization, directed ties. Ties are co-located terrorist attacks by same organisation.
}
\keyword{datasets}
