% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_29}
\alias{covert_29}
\title{Madoff Fraud}
\format{
igraph object
}
\source{
Freely available, to be reconstructed by Manchester http://orgnet.com/Madoff9.png
}
\usage{
covert_29
}
\description{
Bernie Madoff was involved in a massive financial fraud in the USA and was sentenced to 150 years in prison (http://en.wikipedia.org/wiki/Bernard_Madoff). The network is finance flows between financial institutions and Madoff’s firm. All data for this network was gathered from news stories and court documents found on major media web sites. Read more about the social network underpinnings of this scheme in The Network Thinkers blog post (http://www.thenetworkthinkers.com/2009/02/madoff-feeder-funds.html)
1-mode directed network 61 x 61 firm by firm, showing money flows between banks and other organizations, leading ultimately to Madoff’s firm.
Relations are money flows. Arrows show direction of money flow.
}
\keyword{datasets}
