% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_30}
\alias{covert_30}
\title{Madrid Train Bombing 2004}
\format{
list of igraph objects
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks) or here http://doitapps.jjay.cuny.edu/jjatt/data.php
}
\usage{
covert_30
}
\description{
This is a time series that treat specific attacks as endpoints and depict the evolution of relations between individuals indirectly and directly associated with the Madrid train bombing. http://en.wikipedia.org/wiki/2004_Madrid_train_bombings
1-mode stacked matrices 55 x 55 person by person, data on 20 time periods plus kinship data and tie extinguished data.
Codebook available here http://doitapps.jjay.cuny.edu/jjatt/files/Relations_Codebook_Public_Version2.pdf
}
\keyword{datasets}
