% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_31}
\alias{covert_31}
\title{Mali Terrorist Network}
\format{igraph object}
\source{
The dataset has been reconstructed from the following publication: Walther, Olivier J., and Dimitris Christopoulos (2015), "Islamic terrorism and the Malian rebellion." Terrorism and Political Violence, 27 (3), 497-519.
}
\usage{
covert_31
}
\description{
Data refers to a terrorist network operating in the Sahel Sahel-Sahara region and describes relationships between Islamists and Tuareg rebels during the Malian conflict.
Data comes from a selection of newspaper articles published between 2010 and 2012.
“Using social network analysis, our first aim is to illuminate the relationships between the Islamists and the rebels involved in the current Malian conflict. We use a selection of newspaper articles to demonstrate that the connection between Islamists and rebels depends on brokers who passed from the Tuareg rebellion to radical groups. Our second objective is to detail the internal relationships within each of the subgroups. Our findings show how Islamists were affected by the accidental disappearance of one the AQMI regional emirs and how the death of one of the architects of the Tuareg rebellion affected rebel cohesion.” Walther et al., INSNA Sunbelt Conference 2013
1-mode matrix 36x36 person by person
Relations derived from interactions, including participation in political or military event, political meetings; trainings in Afghanistan, Iraq, or Libya; combats; negotiations for hostage releases; or involvements with a killing, an abduction, or a bombing.
}
\references{
Walther, Olivier J., and Dimitris Christopoulos (2015), "Islamic terrorism and the Malian rebellion."
\emph{Terrorism and Political Violence}, 27 (3), 497-519.
}
\keyword{datasets}
