% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_32}
\alias{covert_32}
\title{Montreal Street Gangs}
\format{igraph object}
\source{
The data has been reconstructed from the following article: Karine Descormiers and Carlo Morselli (2011) 'Alliances, Conflicts, and Contradictions in Montreal's Street Gang Landscape' International Criminal Justice Review, Vol. 1 No. 3, pp. 297-314
}
\usage{
covert_32
}
\description{
Data obtained from the Montreal Police’s central intelligence base, the Automated Criminal Intelligence Information System (ACIIS), was used to reconstruct the organization of drug-distribution operations in Montreal North. These operations were targeted during three separate investigations between 2004 and 2007 by the Montreal Police, who believed that the criminal activities were under the control of one of the more reputed gangs in Montreal—the Bo-Gars (or Handsome Boys, in English). Because the trials extending from two of the investigations were still ongoing at the time of analysis, their names remain confidential and I simply refer to Investigations A, B, and C. Investigation A began in February 2004 and ended in April 2005, with the arrests of 27 individuals who were accused primarily of importing and distributing crack and cocaine in a Montreal North neighborhood. Investigation A was the largest of the three investigations under study and it was the only case to offer electronic surveillance information amongst the available data sources. Investigations B and C, which were direct extensions of observations made during Investigation A, both began during the fall of 2006 and ended in June 2007, with the arrests of 24 individuals who were targeted in Investigation B and 11 individuals targeted in Investigation C. Investigation B concentrated on street dealers of marijuana and crack, while Investigation C focused specifically on the activities of a group of wholesalers who were supplying some of the dealers targeted in Investigation B. Overall, 101 individuals were monitored during the investigations—45 in Investigation A, 30 in Investigation B, and 26 in Investigation C. This list of 101 individuals was used as a starting point to reconstruct the final network. This final network was comprised of 70 participants and was based on information obtained from three data sources.
NETWORK  1-mode matrix 35 x 35 organization by organization (Gangs in Montreal). Undirected ties, binary (original network is directed).
\preformatted{
Ties are relationships between gangs (as reported in focus groups/interviews with gang members).
Attribute data:
-- Gang affiliation: 1) Bloods, 2) Crips, or 3) Other
-- Gang Ethnicity: 1) Hispanic, 2) Afro-Canadian, 3) Caucasian, 4) Asian, 5) No main association/mixed.
-- Territory data: 1) Downtown 2) East 3) West
-- Missing data coded as 99
}
}
\references{
Karine Descormiers and Carlo Morselli (2011) 'Alliances, Conflicts, and Contradictions in Montreal's Street Gang Landscape' International Criminal Justice Review, Vol. 1 No. 3, pp. 297-314
}
\keyword{datasets}
