% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_33}
\alias{covert_33}
\title{Ndrangheta Mafia 2}
\format{igraph object}
\source{
The data has been reconstructed by the document "ORDINANZA DI APPLICAZIONE DI MISURA COERCITIVA con mandato di cattura - art. 292 c.p.p. -" which is available online at the following address http://www.stampoantimafioso.it/documentazione-antimafia/ordinanze/.
Stampo Antimafioso is a project which aims to share information about the Mafia operating in Northem Italy.
The dataset has been reconstructed by mostly referring to pp.87-110 of the document named "Operazione Infinito". This report is a judicial document concerning the pre-trial detention order triggered by the the preliminary investigation judge (Giudice per le indagini preliminari) of Milan. With this judicial act, measures of custody and pretrial detention have been ordered for the reported suspected of 'Ndrangheta affiliation.
}
\usage{
covert_33
}
\description{
Data is on attendance of suspected members of the Ndrangheta criminal organization at summits (meetings whose purpose is to make important decisions and/or affiliations, but also to solve internal problems and to establish roles and powers) taking place between 2007 and 2009.
2-mode matrix 156 x 47 persons by events (summits), undirected binary ties.
Attendance at events have been registered by police authorities through wiretapping and observations during the large investigation called "Operazione Infinito".
}
\keyword{datasets}
