% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_34}
\alias{covert_34}
\title{Noordin Top}
\format{list of igraph objects}
\source{
Data is available in its original format from Manchester  (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks).
}
\usage{
covert_34
}
\description{
These
data were drawn primarily from "Terrorism in Indonesia: Noordin's
Networks," a publication of the International Crisis Group (2006) and
include relational data on the 79 individuals listed in Appendix C of that
publication. The data were initially coded by Naval Postgraduate School
students as part of the course “Tracking and Disrupting Dark Networks” under
the direction of Professor Sean Everton, Co-Director of the CORE Lab, and Professor Nancy Roberts. CORE Lab Research
Assistant Daniel Cunningham reviewed and cleaned all coding made by students.
NETWORK  1-mode stacked matrix 79 x 79 person by person. Ties are undirected.
Ties include classmates; friendship; soulmates; co-location of logistical activity; co-attendance at meetings; co-participation in operations; co-attendance at training events; communications; business & financial ties.
Codebook available here http://www.thearda.com/archive/files/codebooks/origCB/Noordin\%20Subset\%20Codebook.pdf
}
\references{
Roberts, Nancy and Sean F. Everton. 2011. Roberts and Everton Terrorist Data: Noordin Top Terrorist Network (Subset).

For a detailed analysis of Noordin top network see: Everton, S. F. (2012) Disrupting dark networks: Cambridge University Press. For a detailed explanation of matrices and the kind of relationship considered see the appendix of the book.
}
\keyword{datasets}
