% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_35}
\alias{covert_35}
\title{Paul Revere}
\format{
list of igraph objects
}
\source{
Freely available: http://kieranhealy.org/blog/archives/2013/06/09/using-metadata-to-find-paul-revere/
}
\usage{
covert_35
}
\description{
The Paul Revere conspiracy dataset concerns relationships between 254 people and their affiliations with seven different organizations in Boston. The dataset refers to Paul Revere, who was responsible for organizing a local militia of Boston's revolutionary movement (see http://en.wikipedia.org/wiki/Sons_of_Liberty). The dataset was analysed by Kieran Healy of Duke University.
This dataset has been reconstructed by looking at the information presented in the appendix of the book ‘Paul Revere's Ride’ published by David Fischer (1994).
2-mode affiliation matrix 254x7 people by organizations, relations refer to membership of organizations; 1-mode matrix 254 x 254 people by people, relations are shared membership of organizations, relations are valued with values indicating number of memberships in common.
}
\references{
Fischer, D. 1994. Paul Revere's ride. Oxford University Press.
}
\keyword{datasets}
