% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_36}
\alias{covert_36}
\title{Philippine Kidnappings}
\format{igraph object}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). http://www.tandfonline.com/eprint/cCV3RJihmG3miPFECpV7/full
}
\usage{
covert_36
}
\description{
Data refers to the Abu Sayyaf Group (ASG), a violent non-state actor operating in the Southern Philippines. In particular, this data is related to the Salast movement that has been founded by Aburajak Janjalani, a native terrorist of the Southern Philippines in 1991. ASG is active in kidnapping and other kinds of terrorist attacks (Gerdes et al. 2014).
The reconstructed 2-mode matrix combines terrorist kidnappers and the terrorist events they have attended.
2-model matrix 246x105 persons by terrorist events, undirected binary relations are participation in events
}
\references{
Gerdes, Luke M., Kristine Ringler, and Barbara Autin. "Assessing the Abu Sayyaf Group's Strategic and Learning Capacities." \emph{Studies in Conflict & Terrorism} 37, no. 3 (2014): 267-293.
}
\keyword{datasets}
