% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_41}
\alias{covert_41}
\title{Siren}
\format{igraph object}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks), reconstructed from Morselli’s book, Inside Criminal Networks http://www.springer.com/social+sciences/criminology/book/978-0-387-09525-7 book page 187, Appendix
}
\usage{
covert_41
}
\description{
Project Siren began in February 1998 when a port worker informed members of the CERVO group that a container of stolen vehicles had been recently shipped to Ghana. This shipment was subsequently seized at its transit point in Anvers, Belgium. This initial tip and action led to a close monitoring of the suspects involved in the shipment. The investigation continued for 4 months (to June 1998), during which time CERVO members monitored stolen-vehicle shipments intended for Russia, Egypt, Iraq, Italy, and Switzerland. Some vehicles were also resold in Toronto. Overall, 35 cars were retrieved according to the files that were consulted.
NETWORK  1-mode matrix 44 x 44 person by person. Ties are undirected.
Relations represent communication exchanges between criminals. Data comes from police wiretapping.
}
\references{
Morselli, C., 2009. Inside criminal networks. New York: Springer.
}
\keyword{datasets}
