% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_43}
\alias{covert_43}
\title{Southeast Asian Aggregated Attacks 2005}
\format{
list of igraph objects
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks) or here http://doitapps.jjay.cuny.edu/jjatt/data.php
}
\usage{
covert_43
}
\description{
The Southeast Asian
Aggregate Attack Series collapses all of the individual, Indonesian cases into
a single series of relations useful for inspecting a series of behavioural and
compositional changes in one terrorist network.
1-mode stacked matrices 109x109 person by person. Kinship, friendship, acquaintanceship, time series, tie formed, tie ended.
Codebook available here http://doitapps.jjay.cuny.edu/jjatt/files/Relations_Codebook_Public_Version2.pdf
}
\keyword{datasets}
