% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_44}
\alias{covert_44}
\title{Suffragettes}
\format{list of igraph objects}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). http://eprints.ncrm.ac.uk/842/1/Social_Network_analysis_Edwards.pdf
}
\usage{
covert_44
}
\description{
Dataset collected by Gemma Edwards on the Suffragette movement in the UK. “…Relational data were constructed from historical archives, including suffragette letters and speeches, and secondary sources like published auto-biographies and newspaper accounts. This historical material provided not only relational data for quantitative network analysis about the structure of these networks, but rich, narrative accounts about the meaning of ties over time and the perception of the network from those within it. Using historical letters as a source of data on suffragette networks was seen as particularly useful for example, as letters contained relational data in terms of ‘who was writing to whom’, and writers would further ‘talk their ties’ within the course of letter writing. Also, letters tend to be dated, allowing for an analysis of the evolution of ties over time (Edwards and Crossley 2009).
\preformatted{
1-mode network 85x85 persons by persons, relations are co-location (1908 visits)
1-mode network 85x85 persons by persons, relations are co-location (1909 visits)
1-mode network 85x85 persons by persons, relations are co-location (1910 visits)
1-mode network 85x85 persons by persons, relations are co-location (1911 visits)
1-mode network 85x85 persons by persons, relations are co-location (1912 visits)
1-mode network 85x85 persons by persons, relations are co-location (1913 visits)
1-mode network 85x85 persons by persons, relations are co-location (Blathwayt visits)
1-mode network 112x112 persons by persons, relations are pre-existing ties (Pankhurst Inner Circle)
2-mode network 13x18 persons by militant acts (Bristol Bath militant acts)
1-mode network 49x49 persons by persons, relations are co-attendence at events (Bristol Bath events)
2-mode network 1216x398 person by arrest date and location (50 all 2M)
}
}
\references{
Edwards, Gemma, and Nick Crossley. "Measures and meanings: exploring the ego-net of Helen Kirkpatrick Watts, militant suffragette." \emph{Methodological Innovations Online} 4.1 (2009): 37-61.

Edwards, Gemma. "Infectious innovations? The diffusion of tactical innovation in social movement networks, the case of suffragette militancy." \emph{Social Movement Studies} 13.1 (2014): 48-69.

Edwards, Gemma. Social Movements and Protest. Cambridge University Press, 2014. Crossley, Nick, et al. "Covert social movement networks and the secrecy-efficiency trade off: The case of the UK suffragettes (1906–1914)." \emph{Social Networks} 34.4 (2012): 634-644.
}
\keyword{datasets}
