% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_45}
\alias{covert_45}
\title{Swingers}
\format{
igraph object
}
\source{
Data from Anne-Marie Niekamp. Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks).
}
\usage{
covert_45
}
\description{
Data on couples attending swinging parties.
2-mode matrix 57 x 39 couples by events (parties)
"Swing units" are a couple attending events with other "swing units".
}
\keyword{datasets}
