% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_46}
\alias{covert_46}
\title{Kenya Tanzania Gerdes}
\format{igraph object}
\source{
Available from Center for Computational Analysis of Social and Organizational Systems (CASOS). (2008). Tanzania-Kenya-imoon.xml. Data available online: http://www.casos.cs.cmu.edu/ computational_tools/datasets/internal/tanzania_ kenya/index11.php.
Also Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks).
}
\usage{
covert_46
}
\description{
Data collected by the Center for Computational Analysis of Social and Organizational Systems, a research group at Carnegie Mellon University, on the participation of 18 Al Qaeda members in 25 functional tasks underlying the 1998 bombings of the U.S. Embassies in Nairobi, Kenya, and Dar es Salaam, Tanzania  2-Mode persons to Standing Committees.
2-mode matrix 18 x 25 persons to tasks, binary undirected. Relations are participation in tasks.
}
\references{
Gerdes, Luke M. (2014), ‘Dependency Centrality from Bipartite Social Networks’, \emph{Connections}, 34, 1&2
}
\keyword{datasets}
