% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-covert.R
\docType{data}
\name{covert_7}
\alias{covert_7}
\title{Bali Bombing 2005}
\format{list of igraph objects}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks) or here http://doitapps.jjay.cuny.edu/jjatt/data.php
}
\usage{
covert_7
}
\description{
This is a time series that treat specific attacks as endpoints and depict the evolution of relations between individuals indirectly and directly associated with the 2005 Bali bombing by Jemaah Islamayah. http://en.wikipedia.org/wiki/2005_Bali_bombings
1-mode matrix 27 x 27 person by person, data for 11 time periods. Ties are undirected and valued. Tie codes: 0 = No relation // 1 = Acquaintances/distant
family ties (interaction limited to radical organisation activities) // 2
= Friends/Moderately close family (inc co-workers/ roommates)
Operational/Org leadership/Operational lies (e.g. worked closely on a
bombing together) // 3 = Close friends/family, tight-knit operational
cliques.
}
\keyword{datasets}
