% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{crime}
\alias{crime}
\title{St. Louis Crimes}
\format{(bipartite) igraph object}
\source{
http://moreno.ss.uci.edu/data#crime
}
\usage{
crime
}
\description{
In the 1990s Rick Rosenfeld and Norm White used police records to collect data on crime in St. Louis. They began with five homicides and recorded the names of all the individuals who had been involved as victims, suspects or witnesses. They then explored the files and recorded all the other crimes in which those same individuals appeared. This snowball process was continued until they had data on 557 crime events. Those events involved 870 participants of which: 569 appeared as victims 682 appeared as suspects 195 appeared as witnesses, and 41 were dual (they were recorded both as victims and suspects in the same crime. Their data appear, then, as an 870 by 557, individual by crime event matrix. Victims are coded as 1, suspects as 2, witnesses as 3 and duals as 4.
}
\keyword{datasets}
