% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defAlly}
\alias{defAlly}
\title{Dyadic data on bilateral alliance}
\format{
A data frame with 2256 rows and 5 variables:
\describe{
  \item{ccode1}{COW country code of the first country}
  \item{ccode2}{COW country code of the second country}
  \item{ij}{dyad id}
  \item{defAlly}{equals 1 if the two countries were allied in that year}
  \item{year}{year of observation}
}
}
\source{
ATOP data
}
\usage{
defAlly
}
\description{
Dyadic data on bilateral alliance
}
\keyword{datasets}
