% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dipl_ties}
\alias{dipl_ties}
\title{A list of 8 time periods of data on diplomatic ties between 134 countries}
\format{
An object of class \code{list} of length 8.
}
\usage{
dipl_ties
}
\description{
A list of 8 time periods of data on diplomatic ties between 134 countries
}
\keyword{datasets}
