% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{dnc_corecipient}
\alias{dnc_corecipient}
\title{DNC Email (Corecipients)}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://www.rene-pickhardt.de/extracting-2-social-network-graphs-from-the-democratic-national-committee-email-corpus-on-wikileaks/
}
\usage{
dnc_corecipient
}
\description{
This is the undirected network of people having received the same email in the 2016 Democratic National Committee email leak.  The Democratic National Committee (DNC) is the formal governing body for the United States Democratic Party.  A dump of emails of the DNC was leaked in 2016, and this dataset contains persons from that dump as nodes, and an edge when two persons received the same email, i.e., when two persons were on the recipient list of the same email.  Multiple edges indicate multiple emails.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.
}
\seealso{
\link{dnc_temporalGraph}
}
\keyword{datasets}
