% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{dolphins_1}
\alias{dolphins_1}
\title{Dolphins (I)}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data.html#dolphin
}
\usage{
dolphins_1
}
\description{
Thirteen male dolphins were observed as they swam in a shallow lagoon. Tabulations were made of who was swimming with whom. The table shows the observed frequencies.
}
\references{
R. C. Connor, R. A. Smolker and A. F. Richards, 1992, "Dolphin alliances and coalitions," in \emph{Coalitions and Alliances in Humans and Other Animals} (Eds: A. H. Harcourt and F. B. M. deWaal). Oxford: Oxford University Press, 415-444.
}
\keyword{datasets}
