% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{dolphins_2}
\alias{dolphins_2}
\title{Dolphins (II)}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data.html#dolphins
}
\usage{
dolphins_2
}
\description{
undirected social network recording frequent associations between pairs in a community of 62 dolphins living off Doubtful Sound, New Zealand.
}
\references{
D. Lusseau, K. Schneider, O. J. Boisseau, P. Haase, E. Slooten, and S. M. Dawson, The bottlenose dolphin community of Doubtful Sound features a large proportion of long-lasting associations, \emph{Behavioral Ecology and Sociobiology} 54, 396-405 (2003).

D. Lusseau, The emergent properties of a dolphin social network, \emph{Proc. R. Soc.} London B (suppl.) 270, S186-S188 (2003).

D. Lusseau, Evidence for social role in a dolphin social network, Preprint q-bio/0607048 (http://arxiv.org/abs/q-bio.PE/0607048)
}
\keyword{datasets}
