% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dwnom}
\alias{dwnom}
\title{Ideological scores of the top 10 most ideological senators in the 109th Congress
#' #' @format A data frame:
\describe{
   \item{labs}{name of the senator}
   \item{dwnom}{dwnominate (ideologocal) score}
}}
\format{
An object of class \code{data.frame} with 10 rows and 2 columns.
}
\usage{
dwnom
}
\description{
Ideological scores of the top 10 most ideological senators in the 109th Congress
#' #' @format A data frame:
\describe{
   \item{labs}{name of the senator}
   \item{dwnom}{dwnominate (ideologocal) score}
}
}
\keyword{datasets}
