% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dyadData}
\alias{dyadData}
\title{A long-form dataset of alliances, contiguity, and war among 50 countries over 10 time periods.}
\format{
An object of class \code{data.frame} with 24500 rows and 10 columns.
}
\usage{
dyadData
}
\description{
A long-form dataset of alliances, contiguity, and war among 50 countries over 10 time periods.
}
\keyword{datasets}
