% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edist}
\alias{edist}
\title{Geographic distances between top 10 ideological senators}
\format{
A 10x10 numeric matrix.
}
\usage{
data(edist)
}
\description{
A 10 by 10 matrix of geographical distances among the top 10 most ideological senators in the 109th Congress.
}
\keyword{datasets}
