% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{eies_messages}
\alias{eies_messages}
\title{EIES (Messages)}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data.html#eies
}
\usage{
eies_messages
}
\description{
These data arose from an early experiment on computer mediated communication. Fifty academics interested in social network research were allowed to contact each other via an Electronic Information Exchange System (EIES). The data collected consisted of all messages sent plus acquaintance relationships at two time periods (collected via a questionnaire).The data include the 32 actors who completed the study. In addition attribute data on primary discipline and number of citations was recorded.

NUMBER_OF MESSAGES is the total number of messages person i sent to j over the entire period of the study.

The attribute data give the number of citations of the actors work in the social science citation index at the beginning of the study together with a discipline code: 1 = Sociology, 2 = Anthropology, 3 = Mathematics/Statistics, 4 = other. These data are used by Wasserman and Faust in their network analysis book.
}
\references{
Freeman, S. C. and L. C. Freeman (1979). The networkers network: A study of the impact of a new communications medium on sociometric structure. \emph{Social Science Research Reports} No 46. Irvine CA, University of California.

Wasserman S. and K. Faust (1994). Social Network Analysis: Methods and Applications.Cambridge University Press, Cambridge.
}
\seealso{
\link{eies_relations}
}
\keyword{datasets}
