% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{el}
\alias{el}
\title{A list of edges-collaborations among the 10 most ideological senators from the 109th Congress.
An edge $ij$ in this network is defined as equal to 1 if i cosponsored j at least two times.
#' @format A data frame:
\describe{
   \item{V1}{name of the first senator}
   \item{V2}{name of the second senator}
}}
\format{
An object of class \code{data.frame} with 27 rows and 2 columns.
}
\usage{
el
}
\description{
A list of edges-collaborations among the 10 most ideological senators from the 109th Congress.
An edge $ij$ in this network is defined as equal to 1 if i cosponsored j at least two times.
#' @format A data frame:
\describe{
   \item{V1}{name of the first senator}
   \item{V2}{name of the second senator}
}
}
\keyword{datasets}
