% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emailnet}
\alias{emailnet}
\title{a statnet network of emails among 54 GAStech employees}
\format{
Two vertices are connected if i sent at least one email to j. Emails are limited to those with fewer than 5 recipients.
\describe{
  \item{name}{employee name}
  \item{position}{employee position in the company}
  \item{mil}{military service branch}
  \item{citizen}{country of citizenship}
  \item{vertex.names}{email address}
}
}
\source{
2014 VAST Challenge
}
\usage{
emailnet
}
\description{
a statnet network of emails among 54 GAStech employees
}
\keyword{datasets}
