% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{euroroad}
\alias{euroroad}
\title{Road network Europe}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://lovro.lpt.fri.uni-lj.si/support.jsp
}
\usage{
euroroad
}
\description{
This is the international E-road network, a road network located mostly in Europe.  The network is undirected; nodes represent cities and an edge between two nodes denotes that they are connected by an E-road.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Lovro Subelj and Marko Bajec. Robust network community detection using balanced propagation. Eur. Phys. J. B, 81(3):353--362, 2011.
}
\keyword{datasets}
