% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{f2f_hypertext}
\alias{f2f_hypertext}
\title{Face-2-face contacts at Hypertext}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://www.sociopatterns.org/
}
\usage{
f2f_hypertext
}
\description{
This is the network of face-to-face contacts of the attendees of the ACM Hypertext 2009 conference. The ACM Conference on Hypertext and Hypermedia 2009 (HT 2009, http://www.ht2009.org/) was held in Turin, Italy over three days from June 29 to July 1, 2009. In the network, a node represents a conference visitor, and an edge represents a face-to-face contact that was active for at least 20 seconds. Multiple edges denote multiple contacts. Each edge is annotated with the time at which the contact took place.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Lorenzo Isella, Juliette Stehlé, Alain Barrat, Ciro Cattuto, Jean-François  Pinton, and Wouter Van den Broeck. What's in a crowd? analysis of face-to-face behavioral networks. J. of Theoretical Biology, 271(1):166--180, 2011.
}
\keyword{datasets}
