% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{f2f_infectious}
\alias{f2f_infectious}
\title{Face-2-face contacts at Infectious}
\format{
igraph object
}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://www.sociopatterns.org/
}
\usage{
f2f_infectious
}
\description{
This network describes the face-to-face behavior of people during the exhibition INFECTIOUS: STAY AWAY in 2009 at the Science Gallery in Dublin. Nodes represent exhibition visitors; edges represent face-to-face contacts that were active for at least 20 seconds. Multiple edges between two nodes are possible and denote multiple contacts. The network contains the data from the day with the most interactions.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Lorenzo Isella, Juliette Stehlé, Alain Barrat, Ciro Cattuto, Jean-François  Pinton, and Wouter Van den Broeck. What's in a crowd? analysis of face-to-face behavioral networks. J. of Theoretical Biology, 271(1):166--180, 2011.
}
\keyword{datasets}
