% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{ffe_influence}
\alias{ffe_influence}
\title{French Financial Elite (influence)}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data#ffe
}
\usage{
ffe_influence
}
\description{
In 1990 Kadushin collected data from 127 members of the French financial elite. He used various criteria to determine the top 28 and recorded their who-to-whom responses to questions about who was influencential, who were members of the elite and who were friends. He also recorded a large amount of information on their individual backgrounds and characteristics.
}
\references{
Kadushin, C. 1995. "Friendship among the French financial elite." \emph{American Sociological Review} 60:202-221.
}
\seealso{
\link{ffe_elite},\link{ffe_friends}
}
\keyword{datasets}
