% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{flo_business}
\alias{flo_business}
\title{Florentine Families (Business)}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data#padgett
}
\usage{
flo_business
}
\description{
Breiger & Pattison (1986), in their discussion of local role analysis, use a subset of data on the social relations among Renaissance Florentine families (person aggregates) collected by John Padgett from historical documents. The two relations are business ties (specifically, recorded financial ties such as loans, credits and joint partnerships) and marriage alliances.

As Breiger & Pattison point out, the original data are symmetrically coded. This is acceptable perhaps for marital ties, but is unfortunate for the financial ties (which are almost certainly directed). To remedy this, the financial ties can be recoded as directed relations using some external measure of power - for instance, a measure of wealth. PADGW provides information on (1) each family's net wealth in 1427 (in thousands of lira); (2) the number of priorates (seats on the civic council) held between 1282- 1344; and (3) the total number of business or marriage ties in the total dataset of 116 families (see Breiger & Pattison (1986), p 239).

Substantively, the data include families who were locked in a struggle for political control of the city of Florence in around 1430. Two factions were dominant in this struggle: one revolved around the infamous Medicis (9), the other around the powerful Strozzis (15).
}
\references{
Breiger R. and Pattison P. (1986). Cumulated social roles: The duality of persons and their algebras. \emph{Social Networks}, 8, 215-256.

Kent D. (1978). The rise of the Medici: Faction in Florence, 1426-1434. Oxford: Oxford University Press.
}
\seealso{
\link{flo_marriage}
}
\keyword{datasets}
