% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gadeData}
\alias{gadeData}
\title{Collaborations among rebel groups}
\format{
A data frame with 930 rows and 11 variables:
\describe{
  \item{Var1}{group name of the first rebel group}
  \item{Var2}{group name of the second rebel group}
  \item{coopActions}{number of cooperative actions between the two groups}
  \item{id}{dyad id}
  \item{ideol_diff.dyad}{ideological distance between the two groups}
  \item{powerdiff.dyad}{power difference between the two groups}
  \item{loc.dyad}{location of the two groups}
  \item{spons.dyad}{common sponsor}
  \item{averageId.node}{first group's ideology}
  \item{size.node}{first group's size}
  \item{spons_actor.node}{first group's sponsor}
}
}
\source{
Gade et al 2019
}
\usage{
gadeData
}
\description{
A dataset to replicate analysis from Gade et al 2019.
}
\keyword{datasets}
