\name{galas}
\alias{galas}
\docType{data}
\title{Galaskiewicz-- CEO's and Clubs}
\description{
galas: One 26 by 15 affiliation matrix.

\bold{BACKGROUND} These data give the affiliation network of 26 CEO's of major corporations and banks and their spouses to 15 clubs, corporate and cultural boards.. Data were collected in the Minneapolis area. Membership was during the period 1978-1981.
}
\usage{data(galas)}

%\details{

%}
\source{
Galaskiewicz J (1985). \emph{Social Organization of an Urban Grants Economy}. New York. Academic Press.
}
%\references{
%}
\examples{
data(galas)

##Plot
plot(galas)



}
\keyword{datasets}
