% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{giraffe}
\alias{giraffe}
\title{Giraffe Affiliation}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data.html#giraffe
}
\usage{
giraffe
}
\description{
The authors studied a herd of six female captive giraffe (Giraffa camelopardalis) for two years. They were concerned with the question of whether giraffe associated randomly or patterned their behavior and proximity in a manner indicative of social relationships. Affiliative interaction, proximity, and nearest neighbors for female giraffe living in a large outdoor enclosure were analyzed, and all three measures were nonrandomly distributed, indicating female giraffe had social preferences. Furthermore, preferences were consistent across measures and time, suggesting that adult female giraffe maintain relationships.
}
\details{
The three different relations (affil, proximity and neighbor) are given in the relation edge attribute.
}
\references{
Brashaw, M. J., M. A. Bloomsmith, T. L. Maple and F. B. Bercovitch. 2007. "The Structure of Social Relationships Among Captive Female Giraffe (Giraffa camelopardalis)." \emph{Journal of Comparative Psychology} 121:46-53.
}
\keyword{datasets}
