% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-misc.R
\docType{data}
\name{glasgow129}
\alias{glasgow129}
\title{Teenage Friends and Lifestyle Study data}
\format{
list of igraph objects
}
\source{
Data downloaded from https://www.stats.ox.ac.uk/~snijders/siena/Glasgow_data.htm
}
\usage{
glasgow129
}
\description{
longitudinal, 3 waves, networks and behavior. For the codebook, see the link provided as source. Only sex.F was recoded to F/M.
}
\keyword{datasets}
