% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-misc.R
\docType{data}
\name{gss_egor}
\alias{gss_egor}
\title{General Social Survey 2004 Egocentric Network Data}
\format{
An egor object (from the egor package) containing three data frames:
\describe{
\item{ego}{594 respondents with the following variables:
\itemize{
\item .egoID: Unique ego identifier
\item vpsu: Variance primary sampling unit
\item vstrat: Variance stratum
\item wtssall: Sample weight
\item age: Respondent's age in years
\item race: Race of respondent
\item sex: Sex of respondent
\item marital: Marital status
\item numgiven: Number of alters named
}
}
\item{alter}{555 alters (network members) with variables:
\itemize{
\item .egoID: Link to ego
\item .alterID: Alter identifier (1-5)
\item age: Alter's age
\item race: Alter's race
\item sex: Alter's sex
\item spouse: Whether alter is spouse/partner
\item intrace: Race of alter (alternative coding)
\item sexsex: Sex of alter (alternative coding)
}
}
\item{aatie}{640 alter-alter ties with variables:
\itemize{
\item .egoID: Link to ego
\item .srcID: Source alter ID
\item .tgtID: Target alter ID
\item weight: Closeness (1 = especially close, 2 = know each other, 3 = total strangers)
}
}
}
}
\source{
General Social Survey 2004. NORC at the University of Chicago.
Data retrieved from \url{https://gss.norc.org/}
}
\usage{
gss_egor
}
\description{
A stratified sample of 594 respondents from the 2004 General Social Survey (GSS) containing egocentric network data. The dataset includes information about respondents (egos), their network members (alters), and the relationships between those network members (alter-alter ties).

The sample was drawn using stratified random sampling without replacement across five stratification variables: age category (3 levels), race, sex, marital status, and number of alters named.
}
\references{
Smith, Tom W., Michael Davern, Jeremy Freese, and Stephen Morgan. General Social Surveys, 1972-2021. Chicago: NORC, 2022.
}
\seealso{
\code{\link[egor]{egor}} for working with egocentric network data
}
\keyword{datasets}
