% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{hens}
\alias{hens}
\title{Hens Pecking order}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data.html#hens
}
\usage{
hens
}
\description{
Records the "peck order" of a flock of 32 White Leghorn hens studied in 1946. A tie from hen a to b means that hen a can peck hen b. The author claims that temporal changes are rare; once a hen dominates another, that pattern persists.
}
\references{
Guhl, A. M., 1953. Social Behavior of the Domestic Fowl. Manhattan, Kansas: Kansas State College, Agricultural Experiment Station, Technical Bulletin 73.
}
\keyword{datasets}
