% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{highlandPonies}
\alias{highlandPonies}
\title{Highland ponies}
\format{
A data matrix with 17 obs. of  18 variables:
\describe{
  \item{ï..}{name of the pony on the row}
}
}
\source{
Shahryar Minhas
}
\usage{
highlandPonies
}
\description{
A dataset on groomin behavior of highland ponies
}
\keyword{datasets}
