% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{highschool_boys}
\alias{highschool_boys}
\title{Friendships among High School Boys}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data.html#high
}
\usage{
highschool_boys
}
\description{
In the fall of 1957. and the spring of 1958. boys in a small high school in Illinois were asked. "What fellows here in school do you go around with most often?" The data are from research reported by Coleman. The data report the direct choices of each of 73 boys at two times. HS1 was recorded in 1957 and HS2 in 1958.
}
\details{
the edge attribute time contains the time period.
}
\references{
Coleman, J. S. Introduction to Mathermatical Sociology. New York: Free Press, pp.450-451.
}
\keyword{datasets}
