% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{ht_advice}
\alias{ht_advice}
\title{High-tech Managers (Advice)}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data#krackht
}
\usage{
ht_advice
}
\description{
These data were collected from the managers of a high-tec company. The company manufactured high-tech equipment on the west coast of the United States and had just over 100 employees with 21 managers. Each manager was asked "To whom do you go to for advice?" and "Who is your friend?" Data for the item "To whom do you report?" was taken from company documents. In addition attribute information was collected. This consisted of the managers age (in years), length of service or tenure (in years), level in the corporate hierarchy (coded 1,2 and 3; 1=CEO, 2 = Vice President, 3 = manager) and department (coded 1,2,3,4 with the CEO in department 0 ie not in a department).
}
\references{
Krackhardt D. (1987). Cognitive social structures. \emph{Social Networks}, 9, 104-134.
}
\seealso{
\link{ht_friends},\link{ht_reports}
}
\keyword{datasets}
