% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{insna}
\alias{insna}
\title{INSNA Teacher Student}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data#ts
}
\usage{
insna
}
\description{
When Barry Wellman founded the International Network for Social Network Analysis (INSNA) in 1977, he sent a questionnaire to all the founding members. Included were questions on who taught each founder and who each founder taught. This data set is based on their responses.
}
\references{
K. Reitz and D. R. White, 1989 "Rethinking the Role Concept: Homomorphisms on Social Networks" pp. 429-488 in L.C.Freeman, D.R. White, A.K.Romney, eds., Research Methods in Social Network Analysis. George Mason Press. Reprinted 1992 Transaction Publishers: New Brunswick, NJ.
}
\keyword{datasets}
