% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interact01}
\alias{interact01}
\title{Interaction: same district and same party in 2001}
\format{
A square binary matrix.
}
\usage{
data(interact01)
}
\description{
A binary matrix indicating whether each legislator pair was from the same district and same party in 2001.
}
\keyword{datasets}
