% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interact03}
\alias{interact03}
\title{Interaction: same district and same party in 2003}
\format{
A square binary matrix.
}
\usage{
data(interact03)
}
\description{
A binary matrix indicating whether each legislator pair was from the same district and same party in 2003.
}
\keyword{datasets}
