% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{jazz}
\alias{jazz}
\title{Collaboration in Jazz}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data.html#jazz
}
\usage{
jazz
}
\description{
The data here record a network of jazz bands. The data were obtained from The Red Hot Jazz Archive digital database. The data include 198 bands that performed between 1912 and 1940, with most of the bands performing in the 1920's. In this case each vertex corresponds to a band, and a link between two bands is established if they had at least one musician in common.
}
\references{
PABLO M. GLEISER and LEON DANON "Community structure in jazz." Advances in Complex Systems (ACS) 2003 Vol: 6 Issue: 4 (December 2003) Page: 565 - 573.
}
\keyword{datasets}
