% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{jpr}
\alias{jpr}
\title{Joint Senate Press Releases}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data.html#jpr
}
\usage{
jpr
}
\description{
These data are from Justin Grimmer's doctoral dissertation in political science at Harvard. They record instances of joint press releases issued by U. S. Senators.
}
\references{
http://people.fas.harvard.edu/~jgrimmer/
}
\keyword{datasets}
