% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{kangaroo}
\alias{kangaroo}
\title{Kangaroo}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data.html#kangaroo
}
\usage{
kangaroo
}
\description{
Frequencies of observed physical proximities among a collection of 17 free-ranging grey kangaroos. Observations were made in the Nadgee Nature Reserve in New South Wales. There were 18 kangaroos in the original report, but one (number M11) was never observed and is therefore dropped from this network.

Two kinds of dominance ranks are included. One, ss, is the ratio of an animal's number of "successes" to its number of "involvements." The other, ps, is calculated by assigning an animal 2 points for each other animal it bests on more than 50\\% of their contacts. One point is given for a tie and none for less than 50\\% successes. Since, except for a juvenile male (M2), there were no cross-sex contests, males and females are ranked seperately, but M2 is ranked with the females.
}
\references{
T. R. Grant, "Dominance and association among members of a captive and a free-ranging group of grey kangaroos (Macropus giganthus)," \emph{Animal Behaviour}, 1973, 21: 449-456.
}
\keyword{datasets}
