% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{karate}
\alias{karate}
\title{Karate Club (binary)}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data#zachary
}
\usage{
karate
}
\description{
These are data collected from the members of a university karate club by Wayne Zachary (presence or absence of ties among the members of the club)

Zachary (1977) used these data and an information flow model of network conflict resolution to explain the split-up of this group following disputes among the members.
}
\references{
Zachary W. (1977). An information flow model for conflict and fission in small groups. \emph{Journal of Anthropological Research}, 33, 452-473.
}
\seealso{
\link{karate_weight}
}
\keyword{datasets}
