% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{karate_weight}
\alias{karate_weight}
\title{Karate Club (weighted)}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data#zachary
}
\usage{
karate_weight
}
\description{
These are data collected from the members of a university karate club by Wayne Zachary (relative strength of the associations, i.e. number of situations in and outside the club in which interactions occurred).

Zachary (1977) used these data and an information flow model of network conflict resolution to explain the split-up of this group following disputes among the members.
}
\references{
Zachary W. (1977). An information flow model for conflict and fission in small groups. \emph{Journal of Anthropological Research}, 33, 452-473.
}
\seealso{
\link{karate}
}
\keyword{datasets}
