% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lazega}
\alias{lazega}
\title{A network of lawyers}
\format{
This data set comes from a network study of corporate law partnership that was
carried out in a Northeastern US corporate law firm, referred to as SG&R, 1988-1991 in New England.
It includes (among others) measurements of networks among the 71 attorneys (partners and associates) of this firm,
i.e. their strong-coworker network, advice network, friendship network, and indirect control networks. Various members' attributes
are also part of the dataset, including seniority, formal status, office in which they work, gender, lawschool attended,
individual performance measurements (hours worked, fees brought in),
attitudes concerning various management policy options, etc.
lazega is a network object compatible with library(network).
This network is directed.  It has the following vertex attributes:
\emph{name}, \emph{Seniority}, \emph{Status} (1 means partner), \emph{Gender} (1 is man, 2 is woman),
 \emph{Office} (1 is Boston, 2 is Hartford, 3 is Providence), \emph{Years} (years with the firm),
\emph{Age}, \emph{Practice} (1 is litigation, 2 is corporate), and
\emph{School} (1 is Harvard or Yale, 2 is University of Connecticut, 3 is other).
 @source Provided to us by Emmanuel Lazega. Please cite the reference below if you use this dataset in your work.
 @references E. Lazega, The Collegial Phenomenon: The Social Mechanisms of
 Cooperation Among Peers in a Corporate Law Partnership. Oxford University Press, Oxford (2001).
}
\usage{
lazega
}
\description{
A network of lawyers
}
\keyword{datasets}
