% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{maayan_faa}
\alias{maayan_faa}
\title{FAA Preferred Routes}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://research.mssm.edu/maayan/datasets/qualitative_networks.shtml
}
\usage{
maayan_faa
}
\description{
This network was constructed from the USA's FAA (Federal Aviation Administration) National Flight Data Center (NFDC), Preferred Routes Database. Nodes in this network represent airports or service centers and links are created from strings of preferred routes recommended by the NFDC.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Federal Aviation Administration. Air traffic control system command center. http://www.fly.faa.gov/.
}
\keyword{datasets}
