% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{maayan_pdzbase}
\alias{maayan_pdzbase}
\title{Maayan Pdzbase}
\format{
igraph object
}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://research.mssm.edu/maayan/datasets/qualitative_networks.shtml
}
\usage{
maayan_pdzbase
}
\description{
This is a network of protein–protein interactions from PDZBase.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Thijs Beuming, Lucy Skrabanek, Masha Y. Niv, Piali Mukherjee, and Harel  Weinstein. PDZBase: A protein--protein interaction database for PDZ-domains. Bioinformatics, 21(6):827--828, 2005.
}
\keyword{datasets}
