\name{mac}
\alias{mac}
\docType{data}
\title{Takahata Macaque Dominance}
\description{
mac: One 62 by 62 matrix of dominance encounters.

\bold{BACKGROUND} The matrix records dominance of the row animal over the column animal in a colony of 62 adult female Japanese macaques (Macaca fuscata fuscata). They are known as the ``Arashiyama B group." Records were made during the non-mating season, April to early October, 1976. Approach-retreat episodes involving food were recorded.

In addition, the presence of six lineages was reported. The first 4 animals belong to a lineage, and the next 14 belong to another. The following 31 are in a third lineage, and the next 6 are in the fourth. The following 6 are the fifth lineage, and the remaining animal is unrelated to the others.
}
\usage{data(mac)}

%\details{
 
%}
\source{
Y. Takahata, \emph{Diachronic changes in the dominance relations of adult female Japanese monkeys of the Arashiyama B group,} in Linda Marie Fedigan and Pamela J. Asquith, eds., The Monkeys of Arashiyama. Albany: State University of New York Press, 1991, pp. 124-139.
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(mac)

##plot
plot(mac)

}
\keyword{datasets}
