% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{macaque}
\alias{macaque}
\title{Macaque Dominance}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data#mac
}
\usage{
macaque
}
\description{
records dominance relations (a directed tie from a to b means a dominates b) in a colony of 62 adult female Japanese macaques (Macaca fuscata fuscata). They are known as the "Arashiyama B group." Records were made during the non-mating season, April to early October, 1976. Approach-retreat episodes involving food were recorded.

In addition, the presence of six lineages was reported. The first 4 animals belong to a lineage, and the next 14 belong to another. The following 31 are in a third lineage, and the next 6 are in the fourth. The following 6 are the fifth lineage, and the remaining animal is unrelated to the others.
}
\references{
Y. Takahata, "Diachronic changes in the dominance relations of adult female Japanese monkeys of the Arashiyama B group," in Linda Marie Fedigan and Pamela J. Asquith, eds., The Monkeys of Arashiyama. Albany: State University of New York Press, 1991, pp. 124-139.
}
\keyword{datasets}
