\name{macaque2}
\alias{macaque2}
\docType{data}
\title{Strayer, Cummins Dominance Among Adult Macaques}
\description{
macaque2: One 30 by 30 matrix of dominance encounters.

\bold{BACKGROUND} This matrix is built from experimental data on the 30 adult members (5 males and 25 females) of a community of macaque monkeys (Macaca mulatta) housed at the Wisconsin Regional Primate Center in Madison. The monkeys were deprived of water and then observations of which monkey got access to a drinking fountain were made.

The data were converted into an index of consistency in linear dominance rankings for six month-long experimental periods. The index is not clearly presented. Individuals are identified with gender tags; Numbers are males, letters are females.
 
}
\usage{data(macaque2)}

%\details{

%}
\source{
F. F. Strayer and Mark S. Cummins, \emph{Aggressive and competitive structures in captive monkey groups,} in Donald R. Omark, F. F. Strayer and Daniel G. Freedman, eds., Dominance Relations: An Ethological View of Human Conflict and Social Interaction. New York: Garland STPM Press, 1980, pp. 85-96.
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(macaque2)

##plot
plot(macaque2)


}
\keyword{datasets}
