% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{mine}
\alias{mine}
\title{Kapferer - Mine}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data#kapmine
}
\usage{
mine
}
\description{
Bruce Kapferer (1969) collected data on men working on the surface in a mining operation in Zambia (then Northern Rhodesia). He wanted to account for the development and resolution of a conflict among the workers. The conflict centered on two men, Abraham and Donald; most workers ended up supporting Abraham.

Kapferer observed and recorded several types of interactions among the workers, including conversation, joking, job assistance, cash assistance and personal assistance. Two miners are connected if they are connected by any of these relations.
}
\references{
Kapferer B. (1969). Norms and the manipulation of relationships in a work context. In J Mitchell (ed), Social networks in urban situations. Manchester: Manchester University Press.

Doreian P. (1974). On the connectivity of social networks. \emph{Journal of Mathematical Sociology}, 3, 245-258.
}
\keyword{datasets}
