% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{miserables}
\alias{miserables}
\title{Les Miserables co-appearances}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data#lesmis
}
\usage{
miserables
}
\description{
Weighted network of co-appearances of characters in Victor Hugo's novel "Les Miserables". Nodes represent characters as indicated by the labels and edges connect any pair of characters that appear in the same chapter of the book. The values on the edges are the number of such coappearances.
}
\references{
D. E. Knuth. (1993). The Stanford GraphBase: A Platform for Combinatorial Computing, Addison-Wesley, Reading, MA
}
\keyword{datasets}
