% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NC01}
\alias{NC01}
\title{Legislator-level covariates in 2001}
\format{
A data frame with one row per legislator.
}
\usage{
data(NC01)
}
\description{
A data frame with node-level covariates for legislators in 2001, including district and party ID.
}
\keyword{datasets}
