% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NC03}
\alias{NC03}
\title{Node-level party and district in 2003}
\format{
An object of class \code{data.frame} with 123 rows and 3 columns.
}
\usage{
NC03
}
\description{
Node-level party and district in 2003
}
\keyword{datasets}
