% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net}
\alias{net}
\title{A 9x9 matrix of edges-collaborations among the 9 most ideological senators from the 109th Congress.
An edge $ij$ in this network is defined as equal to 1 if i cosponsored j at least two times.
#' @format A data frame}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 9 rows and 9 columns.
}
\usage{
net
}
\description{
A 9x9 matrix of edges-collaborations among the 9 most ideological senators from the 109th Congress.
An edge $ij$ in this network is defined as equal to 1 if i cosponsored j at least two times.
#' @format A data frame
}
\keyword{datasets}
