% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkdata.R
\docType{package}
\name{networkdata-package}
\alias{networkdata}
\alias{networkdata-package}
\title{networkdata: A repository of network datasets}
\description{
The package contains a large collection of network datasets with different context.
}
\details{
The package includes data from the following sources:

\itemize{
\item \verb{freeman datasets} from \url{http://moreno.ss.uci.edu/data}
\item \verb{movie networks} from \url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/T4HBA3}
\item \verb{covert networks} from \url{https://sites.google.com/site/ucinetsoftware/datasets/covert-networks}
\item \verb{animal networks} from \url{https://bansallab.github.io/asnr/}
\item \verb{shakespeare networks} build with data from \url{https://github.com/mallaham/Shakespeare-Plays}
\item some networks from \url{http://konect.uni-koblenz.de/}
\item - Tennis networks compiled from \url{https://github.com/JeffSackmann} (please give credit to him if you use this data)
}
and a range of miscellaneous network datasets.

\strong{How to use the datasets}

After loading the \code{networkdata} package, the datasets can
be loaded using the \code{\link{data}} command.

Use \preformatted{  data(package = "networkdata")}
to get the complete list of datasets included in this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schochastics/networkdata}
  \item \url{https://schochastics.github.io/networkdata/}
  \item Report bugs at \url{https://github.com/schochastics/networkdata/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})

}
\keyword{internal}
