% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{petster}
\alias{petster}
\title{Petster Friendships}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
}
\usage{
petster
}
\description{
This Network contains friendships between users of the website hamsterster.com. The network contains many vertex attributes about the pet.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.
}
\keyword{datasets}
