% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{polbooks}
\alias{polbooks}
\title{Political Books}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data#books
}
\usage{
polbooks
}
\description{
Nodes represent books about US politics sold by the online bookseller Amazon.com. Edges represent frequent co-purchasing of books by the same buyers, as indicated by the "customers who bought this book also bought these other books" feature on Amazon.
}
\references{
Valdis Krebs, unpublished, http://www.orgnet.com/
}
\keyword{datasets}
