% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polity}
\alias{polity}
\title{Polity data for 134 countries over 8 years}
\format{
An object of class \code{data.frame} with 1513 rows and 5 columns.
}
\usage{
polity
}
\description{
Polity data for 134 countries over 8 years
}
\keyword{datasets}
