% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{pony}
\alias{pony}
\title{Pony}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data.html#pony
}
\usage{
pony
}
\description{
weights are the number of occasions in which the row pony threatened the column pony.
}
\references{
T.H.Cluton-Brock, J.P.Greenwood and R.P.Powell, 1976, "Ranks and Relationships in Highland Ponies and Highland Cows," \emph{Zeitschrift Tierpsychologie}, 41, 202-216.

M.W.Schein and M.W.Frohman, 1955, "Social Dominance Relationships in a Herd of Dairy-Cattle," \emph{British Journal of Animal Behaviour}, 3, 45-55 (1955).
}
\keyword{datasets}
