% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{powergrid}
\alias{powergrid}
\title{Powergrid}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://toreopsahl.com/datasets/#uspowergrid
}
\usage{
powergrid
}
\description{
This undirected network contains information about the power grid of the Western States of the United States of America. An edge represents a power supply line. A node is either a generator, a transformator or a substation.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Duncan J. Watts and Steven H. Strogatz. Collective dynamics of `small-world' networks. Nature, 393(1):440--442, 1998.
}
\keyword{datasets}
