% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{radoslaw_email}
\alias{radoslaw_email}
\title{Radoslaw - Email Network}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://www.ii.pwr.wroc.pl/~michalski/index.php?content=datasets#manufacturing
}
\usage{
radoslaw_email
}
\description{
This is the internal email communication network between employees of a mid-sized manufacturing company. The network is directed and nodes represent employees. The left node represents the sender and the right node represents the recipient. Edges between two nodes are individual emails.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Radoslaw Michalski, Sebastian Palus, and Przemyslaw Kazienko. Matching organizational structure and social network extracted from  email communication. In Lecture Notes in Business Information Processing, volume 87,  pages 197--206. Springer Berlin Heidelberg, 2011.
}
\keyword{datasets}
