% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{rhesus}
\alias{rhesus}
\title{Rhesus Monkey Grooming}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data#rhesus
}
\usage{
rhesus
}
\description{
observed grooming episodes in a community of free ranging rhesus monkeys in Cayo Santiago observed in June and July of 1963. Seven are males (066, ER, R006, EZ, EC, CY and CN) and the other nine are females.
}
\references{
D. S. Sade, "Sociometrics of Macaca mulatta: Linkages and cliques in grooming matrices," \emph{Folia Primatologica}, 1972, 18: 196-223.
}
\keyword{datasets}
