% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-misc.R
\docType{data}
\name{s50}
\alias{s50}
\title{50-actor excerpt from the Teenage Friends and Lifestyle Study data}
\format{
list of igraph objects
}
\source{
Data downloaded from https://www.stats.ox.ac.uk/~snijders/siena/siena_datasets.htm
}
\usage{
s50
}
\description{
longitudinal, 3 waves, networks and behavior
}
\keyword{datasets}
