% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{same.dist01}
\alias{same.dist01}
\title{Legislators from the same district in 2001}
\format{
A square binary matrix.
}
\usage{
data(same.dist01)
}
\description{
A binary matrix indicating whether each pair of legislators was from the same district in 2001.
}
\keyword{datasets}
