% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{same.dist03}
\alias{same.dist03}
\title{Legislators from the same district in 2003}
\format{
A square binary matrix.
}
\usage{
data(same.dist03)
}
\description{
A binary matrix indicating whether each pair of legislators was from the same district in 2003.
}
\keyword{datasets}
