% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{same.party01}
\alias{same.party01}
\title{Legislators from the same party in 2001}
\format{
A square binary matrix.
}
\usage{
data(same.party01)
}
\description{
A binary matrix indicating whether each pair of legislators belonged to the same political party in 2001.
}
\keyword{datasets}
