% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{same.party03}
\alias{same.party03}
\title{Legislators from the same party in 2003}
\format{
A square binary matrix.
}
\usage{
data(same.party03)
}
\description{
A binary matrix indicating whether each pair of legislators belonged to the same political party in 2003.
}
\keyword{datasets}
