% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_1}
\alias{shakespeare_1}
\title{A Comedy of Errors}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_1
}
\description{
scene co-occurences in Shakespeare's "A Comedy of Errors"
}
\keyword{datasets}
