% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_10}
\alias{shakespeare_10}
\title{Henry IV}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_10
}
\description{
scene co-occurences in Shakespeare's "Henry IV"
}
\keyword{datasets}
