% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_11}
\alias{shakespeare_11}
\title{Henry V}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_11
}
\description{
scene co-occurences in Shakespeare's "Henry V"
}
\keyword{datasets}
