% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_12}
\alias{shakespeare_12}
\title{Henry VI Part 1}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_12
}
\description{
scene co-occurences in Shakespeare's "Henry VI Part 1"
}
\keyword{datasets}
