% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_13}
\alias{shakespeare_13}
\title{Henry VI Part 2}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_13
}
\description{
scene co-occurences in Shakespeare's "Henry VI Part 2"
}
\keyword{datasets}
