% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_15}
\alias{shakespeare_15}
\title{Henry VIII}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_15
}
\description{
scene co-occurences in Shakespeare's "Henry VIII"
}
\keyword{datasets}
