% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_16}
\alias{shakespeare_16}
\title{Julius Caesar}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_16
}
\description{
scene co-occurences in Shakespeare's "Julius Caesar"
}
\keyword{datasets}
