% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_17}
\alias{shakespeare_17}
\title{King John}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_17
}
\description{
scene co-occurences in Shakespeare's "King John"
}
\keyword{datasets}
