% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_18}
\alias{shakespeare_18}
\title{King Lear}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_18
}
\description{
scene co-occurences in Shakespeare's "King Lear"
}
\keyword{datasets}
