% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_19}
\alias{shakespeare_19}
\title{Loves Labours Lost}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_19
}
\description{
scene co-occurences in Shakespeare's "Loves Labours Lost"
}
\keyword{datasets}
