% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_2}
\alias{shakespeare_2}
\title{A Midsummer nights dream}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_2
}
\description{
scene co-occurences in Shakespeare's "A Midsummer nights dream"
}
\keyword{datasets}
