% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_20}
\alias{shakespeare_20}
\title{macbeth}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_20
}
\description{
scene co-occurences in Shakespeare's "macbeth"
}
\keyword{datasets}
