% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_21}
\alias{shakespeare_21}
\title{Measure for measure}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_21
}
\description{
scene co-occurences in Shakespeare's "Measure for measure"
}
\keyword{datasets}
