% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_22}
\alias{shakespeare_22}
\title{Merchant of Venice}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_22
}
\description{
scene co-occurences in Shakespeare's "Merchant of Venice"
}
\keyword{datasets}
