% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_23}
\alias{shakespeare_23}
\title{Merry Wives of Windsor}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_23
}
\description{
scene co-occurences in Shakespeare's "Merry Wives of Windsor"
}
\keyword{datasets}
