% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_25}
\alias{shakespeare_25}
\title{Othello}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_25
}
\description{
scene co-occurences in Shakespeare's "Othello"
}
\keyword{datasets}
