% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_26}
\alias{shakespeare_26}
\title{Pericles}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_26
}
\description{
scene co-occurences in Shakespeare's "Pericles"
}
\keyword{datasets}
