% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_27}
\alias{shakespeare_27}
\title{Richard II}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_27
}
\description{
scene co-occurences in Shakespeare's "Richard II"
}
\keyword{datasets}
