% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_28}
\alias{shakespeare_28}
\title{Richard III}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_28
}
\description{
scene co-occurences in Shakespeare's "Richard III"
}
\keyword{datasets}
