% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_29}
\alias{shakespeare_29}
\title{Romeo and Juliet}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_29
}
\description{
scene co-occurences in Shakespeare's "Romeo and Juliet"
}
\keyword{datasets}
