% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_3}
\alias{shakespeare_3}
\title{A Winters Tale}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_3
}
\description{
scene co-occurences in Shakespeare's "A Winters Tale"
}
\keyword{datasets}
