% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_30}
\alias{shakespeare_30}
\title{Taming of the Shrew}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_30
}
\description{
scene co-occurences in Shakespeare's "Taming of the Shrew"
}
\keyword{datasets}
