% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_31}
\alias{shakespeare_31}
\title{The Tempest}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_31
}
\description{
scene co-occurences in Shakespeare's "The Tempest"
}
\keyword{datasets}
