% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_32}
\alias{shakespeare_32}
\title{Timon of Athens}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_32
}
\description{
scene co-occurences in Shakespeare's "Timon of Athens"
}
\keyword{datasets}
