% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_33}
\alias{shakespeare_33}
\title{Titus Andronicus}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_33
}
\description{
scene co-occurences in Shakespeare's "Titus Andronicus"
}
\keyword{datasets}
