% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_34}
\alias{shakespeare_34}
\title{Troilus and Cressida}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_34
}
\description{
scene co-occurences in Shakespeare's "Troilus and Cressida"
}
\keyword{datasets}
