% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_35}
\alias{shakespeare_35}
\title{Twelfth Night}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_35
}
\description{
scene co-occurences in Shakespeare's "Twelfth Night"
}
\keyword{datasets}
