% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_36}
\alias{shakespeare_36}
\title{Two Gentlemen of Verona}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_36
}
\description{
scene co-occurences in Shakespeare's "Two Gentlemen of Verona"
}
\keyword{datasets}
