% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_4}
\alias{shakespeare_4}
\title{Alls well that ends well}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_4
}
\description{
scene co-occurences in Shakespeare's "Alls well that ends well"
}
\keyword{datasets}
