% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_5}
\alias{shakespeare_5}
\title{Antony and Cleopatra}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_5
}
\description{
scene co-occurences in Shakespeare's "Antony and Cleopatra"
}
\keyword{datasets}
