% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_6}
\alias{shakespeare_6}
\title{As you like it}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_6
}
\description{
scene co-occurences in Shakespeare's "As you like it"
}
\keyword{datasets}
