% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_7}
\alias{shakespeare_7}
\title{Coriolanus}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_7
}
\description{
scene co-occurences in Shakespeare's "Coriolanus"
}
\keyword{datasets}
