% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_8}
\alias{shakespeare_8}
\title{Cymbeline}
\format{igraph object}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_8
}
\description{
scene co-occurences in Shakespeare's "Cymbeline"
}
\keyword{datasets}
