% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shakespeare.R
\docType{data}
\name{shakespeare_9}
\alias{shakespeare_9}
\title{Hamlet}
\format{
igraph object
}
\source{
Raw data downloaded from https://github.com/mallaham/Shakespeare-Plays
}
\usage{
shakespeare_9
}
\description{
scene co-occurences in Shakespeare's "Hamlet"
}
\keyword{datasets}
