% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{sheep}
\alias{sheep}
\title{Bighorn Sheep Dominance}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data.html#sheep
}
\usage{
sheep
}
\description{
Data record wins and losses for 28 female bighorn sheep observed on the National Bison Range in 1984. the weight of a tie from a to b is the number of occasions on which a was observed dominating b. Ages are listed, but those assigned an age of 9 are at least 9 years old; they may be older.
}
\references{
Christine Hass, "Social status in female bighorn sheep (Ovis canadensis): expression, development and reproductive correlates." \emph{Journal of the Zoological Society of London}, 1991, 225: 509-523.Station, Technical Bulletin 73.
}
\keyword{datasets}
