% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{show_networks}
\alias{show_networks}
\title{Helper function to view datasets}
\usage{
show_networks(type = "all")
}
\arguments{
\item{type}{one of "all","two-mode","one-mode","undirected","directed","weighted","unweighted","collection" (all datasets that consist of several networks),"vattr" (all networks containing vertex attributes)}
}
\value{
data.frame of datasets
}
\description{
Browse through the datasets contained in the package by specifying some desired properties.
}
\examples{
# all networks
show_networks(type = "all")

# all directed networks
show_networks(type = "directed")
}
\author{
David Schoch
}
