% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{sn_auth}
\alias{sn_auth}
\title{Social Networks Coauthors}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data#auth
}
\usage{
sn_auth
}
\description{
Chris McCarty prepared a data set for the 2008 INSNA meeting in St. Pete. He recorded all the coauthorships in the Social Networks journal from the beginning to provide a network of networkers. The result was a t-shirt with a graphic design that was sold at the meeting.
After the meeting, Lin Freeman cleaned the data set and made it available on his website. It takes the form of a matrix that records coauthorship among 475 authors who were involved in the production of 295 articles. Cell entries report the number of coaurherships displayed by pairs of authors.
}
\keyword{datasets}
