% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{southern_women}
\alias{southern_women}
\title{Davis - Southern Women}
\format{(bipartite) igraph object}
\source{
http://moreno.ss.uci.edu/data.html#davis
}
\usage{
southern_women
}
\description{
These data were collected by Davis et al. in the 1930s. They represent observed attendance at 14 small social events by 18 Southern women.
}
\references{
Breiger R. (1974). The duality of persons and groups. \emph{Social Forces}, 53, 181-190.

Davis, A. et al. (1941). Deep South. Chicago: University of Chicago Press.
}
\keyword{datasets}
