% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-misc.R
\docType{data}
\name{starwars}
\alias{starwars}
\title{Star Wars Episode 1-7}
\format{
list of igraph objects
}
\source{
Data downloaded from https://github.com/evelinag/StarWars-social-network
}
\usage{
starwars
}
\description{
Scene Co-occurrence of Star Wars Characters (Episode 1-7)
}
\keyword{datasets}
