% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{surfersb}
\alias{surfersb}
\title{Windsurfers (Interactions)}
\format{igraph object}
\source{
http://moreno.ss.uci.edu/data.html#beach
}
\usage{
surfersb
}
\description{
This was a study of windsurfers on a beach in southern California during the fall of 1986. The windsurfing community was fairly clearly divided into at least two sub-communities. Members of each community seemed, to some degree, to limit their interaction to fellow group members. Contacts between members of the two groups occurred, but these were less frequent. Observations of 43 individuals were made for 31 days. All interpersonal contacts among collections of these individuals were recorded.
}
\references{
L. C. Freeman, S. C. Freeman and A. G. Michaelson "On Human Social Intelligence." \emph{Journal of Social and Biological Structures}, 11, 1988, 415-425.

L. C. Freeman, S. C. Freeman and A. G. Michaelson "How Humans See Social Groups: A Test of the Sailer-Gaulin Models." \emph{Journal of Quantitative Anthropology}, 1, 1989, 229-238.
}
\seealso{
\link{surfersc}
}
\keyword{datasets}
