% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{tailor_social}
\alias{tailor_social}
\title{Kapferer - Tailor-Shop (social)}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data#kaptail
}
\usage{
tailor_social
}
\description{
Bruce Kapferer (1972) observed interactions in a tailor shop in Zambia (then Northern Rhodesia) over a period of ten months. His focus was the changing patterns of alliance among workers during extended negotiations for higher wages.

Kapferer recorded two two different types of interaction, recorded at two different times (seven months apart) over a period of one month. This network includes the"sociational" (friendship, socioemotional) interactions.

The data are particularly interesting since an abortive strike occurred after the first set of observations, and a successful strike took place after the second.
}
\references{
Kapferer B. (1972). Strategy and transaction in an African factory. Manchester: Manchester University Press.
}
\seealso{
\link{tailor_work}
}
\keyword{datasets}
