% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{taro}
\alias{taro}
\title{Taro Exchange}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data#taro
}
\usage{
taro
}
\description{
These data represent the relation of gift-giving (taro exchange) among 22 households in a Papuan village. Hage & Harary (1983) used them to illustrate a graph hamiltonian cycle. Schwimmer points out how these ties function to define the appropriate persons to mediate the act of asking for or receiving assistance among group members.
}
\references{
Hage P. and Harary F. (1983). Structural models in anthropology. Cambridge: Cambridge University Press.

Schwimmer E. (1973). Exchange in the social structure of the Orokaiva. New York: St Martins.
}
\keyword{datasets}
