% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRADE}
\alias{TRADE}
\title{A matrix of trade relations in 2014}
\format{
An object of class \code{data.frame} with 195 rows and 195 columns.
}
\source{
https://correlatesofwar.org/ Correlates of War (COW) dyadic trade data version 4.0. See COW manual for details
}
\usage{
TRADE
}
\description{
A matrix of trade relations in 2014
}
\keyword{datasets}
