% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tradenet}
\alias{tradenet}
\title{data on exports between international countries (in millions USD) in 2014}
\format{
An object of class \code{network} of length 5.
}
\usage{
tradenet
}
\description{
data on exports between international countries (in millions USD) in 2014
}
\keyword{datasets}
