% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{ucforum}
\alias{ucforum}
\title{UC forum (2-mode)}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://toreopsahl.com/datasets/#online_forum_network
}
\usage{
ucforum
}
\description{
This bipartite network contains user posts to forums. The users are students at the University of California, Irvine. An edge represents the number of times a (P)erson posted in a (F)orum.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Tore Opsahl and Pietro Panzarasa. Triadic closure in two-mode networks: Redefining the global and local  clustering coefficients. Social Networks, 34, 2011.
}
\seealso{
\link{ucsocial}
}
\keyword{datasets}
