% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{ucsocial}
\alias{ucsocial}
\title{UC forum (messages sent)}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://toreopsahl.com/datasets/#online_social_network
}
\usage{
ucsocial
}
\description{
This directed network contains sent messages between the users of an online community of students from the University of California, Irvine. A node represents a user. A directed edge represents a sent message. Multiple edges denote multiple messages.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Tore Opsahl and Pietro Panzarasa. Clustering in weighted networks. Social Networks, 31(2):155--163, 2009.
}
\seealso{
\link{ucforum}
}
\keyword{datasets}
