% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-misc.R
\docType{data}
\name{us_flights}
\alias{us_flights}
\title{Flights within the US}
\format{
An object of class \code{igraph} of length 276.
}
\usage{
us_flights
}
\description{
A network of flights between US airports
}
\keyword{datasets}
