% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{usa_borders}
\alias{usa_borders}
\title{USA Bordering States}
\format{
igraph object
}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://www-cs-faculty.stanford.edu/~uno/sgb.html
}
\usage{
usa_borders
}
\description{
These are the 48 contiguous states and the District of Columbia of the United States of America (the USA).  They include all states except the states of Alaska and Hawaii, which are not connected by land with the other states, and include the District of Columbia (DC).  An edge denotes that two states share a border.  The US states in the configuration given by this dataset exist since February 14, 1912, when Arizona was admitted as the 48th state, and is current as of 2014.  The states of Alaska and Hawaii were admitted as the 49th and 50th states in 1959, but are not contiguous with the other states, and are not reflected in this dataset.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Donald E. Knuth. The Art of Computer Programming, Volume 4, Fascicle 0:  Introduction to Combinatorial and Boolean Functions. Addison-Wesley, 2008.
}
\keyword{datasets}
