% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-konnect.R
\docType{data}
\name{usflights}
\alias{usflights}
\title{US Flights 2010}
\format{igraph object}
\source{
Data downloaded from http://konect.uni-koblenz.de/
orginaly from http://toreopsahl.com/datasets/#usairports
}
\usage{
usflights
}
\description{
This is the directed network of flights between US airports in 2010.  Each edge represents a connection from one airport to another, and the weight of an edge shows the number of flights on that connection in the given direction, in 2010.
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Web Observatory Workshop, pages 1343-1350, 2013.

Tore Opsahl. Why anchorage is not (that) important: Binary ties and sample  selection, 2011.
}
\keyword{datasets}
