% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{W}
\alias{W}
\title{A 1094 by 1094 connectivity matrix of ideological distance between pairs of countries. Larger values correspond to more dissimilar countries.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1094 rows and 1094 columns.
}
\usage{
W
}
\description{
A 1094 by 1094 connectivity matrix of ideological distance between pairs of countries. Larger values correspond to more dissimilar countries.
}
\keyword{datasets}
