% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{war}
\alias{war}
\title{A list of 10 time periods of data on war between 50 countries}
\format{
An object of class \code{list} of length 10.
}
\usage{
war
}
\description{
A list of 10 time periods of data on war between 50 countries
}
\keyword{datasets}
