% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-freeman.R
\docType{data}
\name{wiring}
\alias{wiring}
\title{Bank Wiring Room}
\format{
igraph object
}
\source{
http://moreno.ss.uci.edu/data#wiring
}
\usage{
wiring
}
\description{
These are the observational data on 14 Western Electric (Hawthorne Plant) employees from the bank wiring room first presented in Roethlisberger & Dickson (1939). The data are better known through a scrutiny made of the interactions in Homans (1950), and the CONCOR analyses presented in Breiger et al (1975).

The employees worked in a single room and include two inspectors (I1 and I3), three solderers (S1, S2 and S3), and nine wiremen or assemblers (W1 to W9). The interaction categories include: RDGAM, participation in horseplay; RDCON, participation in arguments about open windows; RDPOS, friendship; RDNEG, antagonistic (negative) behavior; RDHLP, helping others with work; and RDJOB, the number of times workers traded job assignments.

\strong{The dataset only includes the positive and negative ties, making it a signed network}
}
\references{
Breiger R., Boorman S. and Arabie P. (1975). An algorithm for clustering relational data with applications to social network analysis and comparison with multidimensional scaling. \emph{Journal of Mathematical Psychology}, 12, 328-383.

Homans G. (1950). The human group. New York: Harcourt-Brace.

Roethlisberger F. and Dickson W. (1939). Management and the worker. Cambridge: Cambridge University Press.
}
\keyword{datasets}
