% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wmat}
\alias{Wmat}
\title{A 42x42 binary connectivity matrix. This is a toy dataset.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 42 rows and 42 columns.
}
\usage{
Wmat
}
\description{
A 42x42 binary connectivity matrix. This is a toy dataset.
}
\keyword{datasets}
