% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y}
\alias{Y}
\title{A toy vector with 42 observations of a binary variable.}
\format{
An object of class \code{numeric} of length 42.
}
\usage{
Y
}
\description{
A toy vector with 42 observations of a binary variable.
}
\keyword{datasets}
