% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMnet01}
\alias{AMnet01}
\title{Co-sponsorship network pre-redistricting}
\format{
An object of class \code{network} of length 5.
}
\usage{
AMnet01
}
\description{
Co-sponsorship network pre-redistricting
}
\keyword{datasets}
