% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMnet03}
\alias{AMnet03}
\title{Co-sponsorship network post-redistricting}
\format{
A square binary matrix.
}
\usage{
data(AMnet03)
}
\description{
A binary adjacency matrix representing co-sponsorship ties among legislators in 2003, after redistricting.
}
\keyword{datasets}
