% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edist}
\alias{edist}
\title{A 10 by 10 matrix of geograohical distances among the top 10 most ideological senators in the 109th Congress}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 10 columns.
}
\usage{
edist
}
\description{
A 10 by 10 matrix of geograohical distances among the top 10 most ideological senators in the 109th Congress
}
\keyword{datasets}
