% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interact01}
\alias{interact01}
\title{Interaction between same district and same party in 2001}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 123 rows and 123 columns.
}
\usage{
interact01
}
\description{
Interaction between same district and same party in 2001
}
\keyword{datasets}
