% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NC03}
\alias{NC03}
\title{Legislator-level covariates in 2003}
\format{
A data frame with one row per legislator.
}
\usage{
data(NC03)
}
\description{
A data frame with node-level covariates for legislators in 2003, including district and party ID.
}
\keyword{datasets}
