% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{same.dist01}
\alias{same.dist01}
\title{Whether two legistlators are from the same district in 2001}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 123 rows and 123 columns.
}
\usage{
same.dist01
}
\description{
Whether two legistlators are from the same district in 2001
}
\keyword{datasets}
