% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{same.party03}
\alias{same.party03}
\title{Whether two legistlators are from the same party in 2003}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 123 rows and 123 columns.
}
\usage{
same.party03
}
\description{
Whether two legistlators are from the same party in 2003
}
\keyword{datasets}
