% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trade}
\alias{trade}
\title{Correlates of War (COW) dyadic trade data version 4.0. See COW manual for details}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 886828 rows and 23 columns.
}
\source{
https://correlatesofwar.org/
}
\usage{
trade
}
\description{
Correlates of War (COW) dyadic trade data version 4.0. See COW manual for details
}
\keyword{datasets}
