% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{W}
\alias{W}
\title{A 100x100 binary connectivity matrix. This is a toy dataset.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 100 columns.
}
\usage{
W
}
\description{
A 100x100 binary connectivity matrix. This is a toy dataset.
}
\keyword{datasets}
